% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov.R
\name{AOV_DEXi}
\alias{AOV_DEXi}
\title{Calculate AOV results}
\usage{
AOV_DEXi(aTree)
}
\arguments{
\item{aTree}{A decision tree object to perform the analysis on.}
}
\value{
A list with the results of the AOV.
}
\description{
Calculates the results of an Analysis of Variance (AOV).
}
