% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov.R
\name{aov_tree}
\alias{aov_tree}
\title{Calculate AOV Results}
\usage{
aov_tree(tree)
}
\arguments{
\item{tree}{Tree object for analysis.}
}
\value{
A list containing AOV results.
}
\description{
Provides the results of an Analysis of Variance (AOV) based on a given tree.
}
