% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree.R
\name{are_leaves_unevaluated}
\alias{are_leaves_unevaluated}
\title{Check if Any Leaves are Unevaluated}
\usage{
are_leaves_unevaluated(leaves, results, nodes)
}
\arguments{
\item{leaves}{A \code{character} vector listing the leaves to be checked.}

\item{results}{A named \code{numeric} vector, where names correspond to
\code{Node} names and values indicate the evaluation status of the node. A
value of -1 implies the \code{Node} is unevaluated.}

\item{nodes}{A list of \code{Node}s, where each entry contains details of a
node.}
}
\value{
A \code{logical}. Returns \code{TRUE} if any of the leaves are yet to be
evaluated; otherwise, it returns \code{FALSE}.
}
\description{
Determines if any of the provided leaves within a \code{Tree} have not yet
been evaluated. A leaf is deemed unevaluated if its corresponding value in
the results vector is -1.
}
\seealso{
Related functions and objects that might be of interest:
\itemize{
\item \code{\link{Tree-class}}: For a more detailed understanding of the
\code{Tree} class.
\item \code{\link{Node-class}}: To delve deeper into the structure of a
\code{Node}.
\item \code{\link{evaluate_order}}: A function that determines the sequence
for evaluating \code{Node}s in a \code{Tree}.
}
}
\keyword{internal}
