% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_index.R
\name{calculate_conditional_probabilities}
\alias{calculate_conditional_probabilities}
\title{Calculate Conditional Probabilities}
\usage{
calculate_conditional_probabilities(input_table, weight_list, sy, y_levels)
}
\arguments{
\item{input_table}{A \code{matrix} containing all level combinations of the
\( A_i \) factors with the associated \( Y \) values in the last column.}

\item{weight_list}{A \code{list} with the weight vectors for each \( A_i \)
variable. If missing, levels of \( A_i \) have equal weights.}

\item{sy}{A \code{numeric} indicating the number of unique \( Y \) values.}

\item{y_levels}{Optionally, a \code{vector} specifying the \( Y \) levels. If
missing, they are extracted from the \code{input_table}.}
}
\value{
A \code{list} comprising matrices of \( Y \) probabilities
conditional to each \( A_i \) factor. The last element is the vector of
marginal \( Y \) probabilities.
}
\description{
Determines the probabilities of \( Y \) given its direct descendants \( A_i
\).
}
\keyword{internal}
