% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{calculate_digit}
\alias{calculate_digit}
\title{Calculate Digit}
\usage{
calculate_digit(number_of_attributes)
}
\arguments{
\item{number_of_attributes}{A non-negative number representing the number of
attributes.}
}
\value{
The number of digits in the \code{number_of_attributes}. Returns 1 if
\code{number_of_attributes} is 0.
}
\description{
This is an internal helper function used within the package.
It's used in the \code{show} method for the \code{Tree} class to calculate the number
of digits in \code{number_of_attributes}.
}
\details{
Calculates the number of digits of a given non-negative number. Returns 1 if
the input is 0.
}
\keyword{internal}
