% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_index.R
\name{calculate_sensitivity_indices}
\alias{calculate_sensitivity_indices}
\title{Calculate Sensitivity indices (SI) From Conditional Probabilities}
\usage{
calculate_sensitivity_indices(conditional_prob_list, weight_list)
}
\arguments{
\item{conditional_prob_list}{List of matrices of conditional probabilities
(Y conditional to each A_i) plus the vector of Y probabilities.}

\item{weight_list}{The list of weights for the A_i factor levels.}
}
\value{
A vector of Sensitivity Indices (SI).
}
\description{
Calculates the first-order sensitivity indices of Y concerning A_i descendants,
given the conditional probabilities and A_i weights.
}
