% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensitivityIndex.R
\name{clcSI_DEXi}
\alias{clcSI_DEXi}
\title{Title}
\usage{
clcSI_DEXi(aTree, avoidrep = F)
}
\arguments{
\item{aTree}{aTree}

\item{avoidrep}{T if you want to avoid repeted nodes}
}
\description{
Title
}
