% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Options.R
\name{compareScenario}
\alias{compareScenario}
\title{Title}
\usage{
compareScenario(aTree, theScenarios, listNodes)
}
\arguments{
\item{aTree}{aTree}

\item{theScenarios}{theScenarios}

\item{listNodes}{listNodes}
}
\description{
Title
}
