% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option.R
\name{compute_aggregated_values}
\alias{compute_aggregated_values}
\title{Calculate Values for Leaf-Aggregated Scenarios}
\usage{
compute_aggregated_values(tree, results)
}
\arguments{
\item{tree}{A \code{Tree} object.}

\item{results}{A \code{numeric} vector that serves as a placeholder for the
evaluated values.}
}
\value{
A \code{numeric} vector with computed aggregated values.
}
\description{
Computes aggregated values for leaf-aggregated scenarios based on the tree's
evaluation order. Ensures that values are calculated in the correct
sequence to account for aggregation.
}
\keyword{internal}
