% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov.R
\name{compute_aov_sensitivity_effects}
\alias{compute_aov_sensitivity_effects}
\title{Compute Sensitivity Criteria for Model Terms}
\usage{
compute_aov_sensitivity_effects(aov_obj)
}
\arguments{
\item{aov_obj}{An \code{aov} object resulting from a call to the \code{aov()}
function.}
}
\value{
A \code{data.frame} detailing the sensitivity criteria for each model
term. The data frame consists of the following columns:
\itemize{
\item \code{df}: degrees of freedom.
\item \code{ss}: sum of squares.
\item \code{ss.ratio}: ratio of the term's sum of squares to the model's
total sum of squares.
\item \code{cm}: mean square (or variance) associated with each term.
\item \code{F}: F-statistic relevant to each term.
}
Entries are organized in descending order based on the \code{ss.ratio}
values.
}
\description{
Derives sensitivity criteria for each term present in a provided fitted
model.
}
\keyword{internal}
