% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov.R
\name{compute_aov_total_sensitivity}
\alias{compute_aov_total_sensitivity}
\title{Calculate total Sensitivity Factors for Model Terms}
\usage{
compute_aov_total_sensitivity(aov_obj)
}
\arguments{
\item{aov_obj}{An object of class \code{aov} from a call to \code{aov()}.}
}
\value{
A data frame with total sensitivity factors for each term in the
model.
}
\description{
Computes total sensitivity factors for each term in a fitted model.
}
