% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option.R
\name{compute_final_aggregated_values}
\alias{compute_final_aggregated_values}
\title{Calculate Final Aggregated Values}
\usage{
compute_final_aggregated_values(tree, results)
}
\arguments{
\item{tree}{A \code{Tree} object.}

\item{results}{A \code{numeric} vector serving as a placeholder for the
evaluated values.}
}
\value{
A \code{numeric} vector with the final computed aggregated values.
}
\description{
Computes the final aggregated values using the tree's structure. Ensures that
values are propagated correctly through the tree.
}
\keyword{internal}
