% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_index.R
\name{compute_indirect_probabilities}
\alias{compute_indirect_probabilities}
\title{Compute Indirect Probabilities}
\usage{
compute_indirect_probabilities(
  node,
  probs,
  conditional_probabilities,
  tree,
  avoid_repetition
)
}
\arguments{
\item{node}{The specific node for which the indirect probabilities should be
calculated.}

\item{probs}{The computed probabilities for the node.}

\item{conditional_probabilities}{The existing conditional probabilities.}

\item{tree}{A decision tree object.}

\item{avoid_repetition}{A boolean determining if repeated nodes should be
avoided.}
}
\value{
A list of computed indirect probabilities.
}
\description{
Computes the indirect probabilities for descendants of a node based on its
children, computed probabilities, and the existing conditional probabilities.
}
