% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{compute_leaf_weights}
\alias{compute_leaf_weights}
\title{Estimated Weights for Node}
\usage{
compute_leaf_weights(node)
}
\arguments{
\item{node}{A \code{Node} object.}
}
\value{
A numeric vector representing estimated weights for each leaf.
}
\description{
Computes weights for each leaf of a given \code{Node} object, using a linear
regression model. In this model, the root node's aggregation table values
serve as the response variable, and the leaf values as the predictors.
}
\details{
Once derived, coefficients (excluding the intercept) represent the relative
contributions of each leaf towards the root node. By normalizing these
coefficients such that their sum is 1, we obtain an estimate for each leaf's
weight.
}
\seealso{
\itemize{
\item \code{\link{Node-class}}: For more details on the Node class.
}
}
\keyword{internal}
