% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option.R
\name{compute_values_from_aggregation_table}
\alias{compute_values_from_aggregation_table}
\title{Compute Node Values from Aggregation Table}
\usage{
compute_values_from_aggregation_table(tree, results)
}
\arguments{
\item{tree}{A Tree object.}

\item{results}{Numeric vector with pre-existing values.}
}
\value{
Numeric vector with updated node values.
}
\description{
Takes a tree and pre-existing results to determine node values.
Iteratively adjusts values based on the aggregation table and related children.
}
