% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensitivityIndex.R
\name{condprob.direct}
\alias{condprob.direct}
\title{Title}
\usage{
condprob.direct(table, weightlist, sy, Ylevels)
}
\arguments{
\item{table}{a matrix giving all the level combinations of the A_i factors
and, in the last column, the associated Y values}

\item{weightlist}{a list whose elements are the weight vectors of each A_i
variable. If missing, A_i levels are assumed to have equal weights}

\item{sy}{sy}

\item{Ylevels}{optional argument giving the Y levels. If missing, the Ylevels
are extracted from the table}
}
\description{
Calculates the probabilities of Y=y and the probabilities of Y conditional
to its direct descendants A_i, when given the complete table of the
Y modalities with respect to the A_i factors, and when given the A_i
probabilities
}
