% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree.R
\name{convertTreeClass}
\alias{convertTreeClass}
\title{Convert old Tree class objects to new Tree class objects}
\usage{
convertTreeClass(oldTree)
}
\arguments{
\item{oldTree}{An object of the old "Tree" class. This should be a valid
"Tree" object.}
}
\value{
A new object of the new "Tree" class, with the attribute names
changed.
}
\description{
This function takes an object of the old "Tree" class (from JEB's scripts)
and returns a new object of the new "Tree" class (package's one), with the
attribute names changed according to the new class definition.
}
