% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree.R
\name{correct_apost}
\alias{correct_apost}
\title{Correct apostroph in Tree before creation}
\usage{
correct_apost(xml_tree)
}
\arguments{
\item{xml_tree}{the xml to be corrected}
}
\value{
the corrected xml
}
\description{
Correct apostroph in Tree before creation
}
\keyword{internal}
