% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{count_digits}
\alias{count_digits}
\title{Determine Number of Digits in a Non-negative Number}
\usage{
count_digits(number_of_attributes)
}
\arguments{
\item{number_of_attributes}{A non-negative number for which the number of
digits is required.}
}
\value{
An integer representing the number of digits in the
\code{number_of_attributes}.
}
\description{
Computes the digit count for a provided non-negative number, returning 1 if
the value is 0. Used as an auxiliary function within the package, especially
for the \code{show} method of the \code{Tree} class to determine the number of digits
in \code{number_of_attributes}.
}
\details{
If the input \code{number_of_attributes} is 0, the function returns 1.
Otherwise, it computes the number of digits using logarithmic operations.
}
\keyword{internal}
