% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{create_aggregation_matrix}
\alias{create_aggregation_matrix}
\title{Create Aggregation Matrix for Node}
\usage{
create_aggregation_matrix(
  node,
  expected_weights,
  number_of_tables = 1,
  population_size = 50,
  iterations = 50
)
}
\arguments{
\item{node}{A Node object.}

\item{expected_weights}{Numeric vector of expected weights.}

\item{number_of_tables}{Numeric, number of tables (default is 1).}

\item{population_size}{Numeric, population size for genetic algorithm
(default is 50).}

\item{iterations}{Numeric, number of iterations for genetic algorithm
(default is 50).}
}
\value{
A list of aggregation matrices.
}
\description{
Creates an aggregation matrix for a node using genetic algorithm
optimization.
}
\details{
Utilize a genetic algorithm to determine the optimal aggregation matrix.
Here, the "gene" corresponds to the value of the aggregated node. The
optimization function aims to minimize the squared difference between the
expected weight and the actual weight, which ensures that the weights closely
match their expected values.
}
