% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov.R
\name{create_aov_barplot}
\alias{create_aov_barplot}
\title{AOV Bar Plot Generator}
\usage{
create_aov_barplot(
  data,
  is_effect,
  horizontal,
  num_plots,
  show_main,
  axis_label_style,
  ...
)
}
\arguments{
\item{data}{A \code{data.frame} containing the AOV results designated for
plotting.}

\item{is_effect}{\code{logical} signifying whether the provided data
showcases effects (if \code{TRUE}) or totals (if \code{FALSE}).}

\item{horizontal}{\code{logical} indicating the orientation of the bar plots.
If \code{TRUE}, plots are displayed horizontally; otherwise, vertically.}

\item{num_plots}{\code{numeric} dictating the count of effects to be
presented in the plot.}

\item{show_main}{\code{logical} stating if the main effects and total sums of
squares should be amalgamated and exhibited. Relevant only when \code{is_effect}
is \code{FALSE}.}

\item{axis_label_style}{\code{numeric} defining the label styling for the
plot axes.}

\item{...}{Supplementary arguments relayed to the \code{barplot} function for
aesthetic adjustments.}
}
\description{
Constructs bar plots to illustrate the outcomes of an Analysis of Variance
(AOV). Tailored to visualize both the total sum of squares and specific
effects.
}
\keyword{internal}
