% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synoptique.R
\name{create_data_frame_for_plotting}
\alias{create_data_frame_for_plotting}
\title{Produce Dataframe Suitable for Plotting from Tree Structure}
\usage{
create_data_frame_for_plotting(tree)
}
\arguments{
\item{tree}{\code{Tree} structure object.}
}
\value{
\code{data.frame} that is structured for plotting.
}
\description{
Transforms a given tree structure into a dataframe, which is organized
and formatted in a manner that makes it suitable for plotting purposes.
}
\keyword{internal}
