% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov.R
\name{create_factorial_plan}
\alias{create_factorial_plan}
\title{Factorial Plan Generation for a Decision Tree}
\usage{
create_factorial_plan(tree)
}
\arguments{
\item{tree}{\code{Tree} object to be analyzed.}
}
\value{
A \code{matrix} detailing the generated factorial plan.
}
\description{
Extracts the leaf indices from a provided decision tree and constructs a
factorial plan utilizing the \code{gen.factorial} function.
}
\keyword{internal}
