% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree.R
\name{create_node}
\alias{create_node}
\title{Create a Node for the Decision Tree}
\usage{
create_node(node_path, main_tree)
}
\arguments{
\item{node_path}{A character vector representing the path to the node.}

\item{main_tree}{An XML object representing the main tree structure
containing the XML data.}
}
\value{
A new "Node" object containing the following components:
\describe{
\item{Name}{The name of the node.}
\item{Depth}{The depth of the node within the tree.}
\item{Twin}{Numeric value representing a twin node, initialized to 0.}
\item{IsLeaf}{Boolean indicating if the node is a leaf.}
\item{IsLeafAndAggregated}{Boolean, initialized to FALSE.}
\item{Mother}{Character representing the mother of the node.}
\item{Sisters}{Character vector representing the sisters of the node.}
\item{Children}{Character vector representing the children of the node.}
\item{Aggregation}{Aggregation structure for the node.}
\item{RangeScale}{The range scale of the node.}
\item{ScaleLabel}{Character vector representing the scale label of the
node.}
\item{Probability}{Numeric vector representing the probability or weights
for the node.}
\item{NodePath}{Character vector representing the path to the node.}
}
}
\description{
Constructs a new node for the decision tree, incorporating various attributes
and details including leaf status, children, mother, sisters, scale, scale
label, aggregation, and weight list.
}
