% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option.R
\name{create_options}
\alias{create_options}
\title{Generate Random Options Matrix for a Given Tree}
\usage{
create_options(tree, num_options = 1, seed = NULL)
}
\arguments{
\item{tree}{Tree object.}

\item{num_options}{integer. Number of options to generate. Default is 1.}

\item{seed}{integer. Seed for random number generation. Default is NULL (no
seed set).}
}
\value{
A matrix where rows represent tree leaves and columns represent
sampled options.
}
\description{
Create random options based on the \code{RangeScale} and \code{Probability} attributes
of tree nodes. This is useful for creating random scenarios for simulation or
analysis.
}
