% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synoptique.R
\name{create_plot}
\alias{create_plot}
\title{Create Synoptic Plot from Data Frame}
\usage{
create_plot(df2)
}
\arguments{
\item{df2}{\code{data.frame} containing plotting details, which should
include coordinates for boxes, normalized evaluation values, and labels.}
}
\value{
A \code{ggplot} object depicting the synoptic plot of the decision
tree.
}
\description{
Constructs a synoptic plot visualizing nodes of a decision tree based on the
plotting details provided in a data frame. The plot is color-coded based on
normalized evaluations, with rectangles representing tree nodes and labels
indicating node attributes.
}
\keyword{internal}
