% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree.R
\name{create_tree}
\alias{create_tree}
\title{Create a Tree from DEXi's XML output}
\usage{
create_tree(main_tree)
}
\arguments{
\item{main_tree}{An XML object representing the main tree structure. It must
be a valid XML object containing the DEXi decision model.}
}
\value{
A list of tree objects, each representing a distinct root attribute
within the DEXi XML structure. Each tree object contains details about the
root attribute, nodes, leaves, depth, and other related information
regarding the structure of the decision model.
}
\description{
Takes the XML output from a DEXi model and constructs a tree structure.
The function iterates through each root attribute in the XML and constructs
tree objects representing the hierarchical decision model.
}
\seealso{
\code{\link{process_aggregated_leaf}},
\code{\link{process_duplicated_leaf}},
\code{\link{get_dexi_attributes}},
\code{\link{get_paths}}
}
