% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{createAggregationMatrix}
\alias{createAggregationMatrix}
\alias{createAggregationMatrix.Node}
\title{Create Aggregation Matrix for Node}
\usage{
createAggregationMatrix(
  aNode,
  expectedWeight,
  nbTables = 1,
  popSize = 50,
  iters = 50
)
}
\arguments{
\item{aNode}{A Node object.}

\item{expectedWeight}{Numeric vector of expected weights.}

\item{nbTables}{Numeric, number of tables (default is 1).}

\item{popSize}{Numeric, population size for genetic algorithm (default is
50).}

\item{iters}{Numeric, number of iterations for genetic algorithm (default is
50).}
}
\value{
A list of aggregation matrices.
}
\description{
Creates an aggregation matrix for a node using genetic algorithm
optimization.
}
