% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree.R
\name{createNode}
\alias{createNode}
\title{Create a node for the decision tree}
\usage{
createNode(listeNoeuds, MT)
}
\arguments{
\item{listeNoeuds}{A list of nodes}

\item{MT}{The main tree (an XML object)}
}
\value{
A new Node object
}
\description{
Create a node for the decision tree
}
