% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option.R
\name{createOptions}
\alias{createOptions}
\title{Generate a matrix of random options for a given tree}
\usage{
createOptions(aTree, nbOptions = 1, aSeed = NULL)
}
\arguments{
\item{aTree}{An object of class Tree.}

\item{nbOptions}{n integer representing the number of options to generate.
Default is 1.}

\item{aSeed}{An integer seed to be used for random number generation. Default
is NULL, which means no seed is set.}
}
\value{
A matrix of size \code{aTree@NumberOfLeaves} x \code{nbOptions}, where each row
represents a leaf of the tree and each column represents a sampled option.
}
\description{
The options are generated based on the \code{rangeScale} and \code{Proba} properties of
the tree nodes. This is useful for creating random scenarios for simulation
or analysis.
}
