% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_subtree.R
\name{createSubTree}
\alias{createSubTree}
\title{Title}
\usage{
createSubTree(aTree, nodeName, avoidrep = F)
}
\arguments{
\item{aTree}{aTree}

\item{nodeName}{nodeName}

\item{avoidrep}{T if you want to avoid repeted nodes}
}
\description{
Title
}
