% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth_order.R
\name{depth_order}
\alias{depth_order}
\title{Get Aggregated Attributes Ordered by Depth}
\usage{
depth_order(tree)
}
\arguments{
\item{tree}{A decision tree object.}
}
\value{
A vector of aggregated attribute names, ordered by depth.
}
\description{
Returns aggregated attributes ordered by depth. Typically
used for analysis strategy (AS).
}
