% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{describe,Tree-method}
\alias{describe,Tree-method}
\title{describeTree}
\usage{
\S4method{describe}{Tree}(object)
}
\arguments{
\item{object}{Object}

\item{Tree}{The Tree object to be described.}
}
\description{
A method to describe the structure of a Tree object in a comprehensive way.
It prints each node and its properties separately.
}
