% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{describe}
\alias{describe}
\title{Generic Describe Function for Objects}
\usage{
describe(object)
}
\arguments{
\item{object}{The object to be described.}
}
\value{
No return value, called for the side effect of a description of the object
(description of each node in case of a Tree object)
}
\description{
Provides a comprehensive description of an object.
}
\seealso{
\itemize{
\item \code{\link{Tree-class}}: For the Tree class definition.
\item \code{\link{print.Tree}}: For printing a Tree object.
\item \code{\link{show.Tree}}: For showing a Tree object.
}
}
