% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree.R
\name{determine_aggregation}
\alias{determine_aggregation}
\title{Determine Node Aggregation}
\usage{
determine_aggregation(node_path, main_tree, is_leaf, children)
}
\arguments{
\item{node_path}{\code{character} vector representing the path to the node.}

\item{main_tree}{Object containing the XML tree structure data.}

\item{is_leaf}{\code{logical} indicating if the node is a leaf.}

\item{children}{\code{list} of node's children.}
}
\value{
A matrix representing the aggregation structure. If not a leaf, it
includes the factorial plan and transformed 'LOW' function values. For leaf
nodes, the matrix contains zero.
}
\description{
Calculates the aggregation structure of a node. Essential for understanding
how information combines at the node level. Retrieves 'LOW' function values
from XML, transforms them, and determines scales using previous nodes.
Computes a factorial plan for nodes with multiple children, otherwise uses a
sequence.
}
\keyword{internal}
