% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree.R
\name{determine_children}
\alias{determine_children}
\title{Find the Children of a Node}
\usage{
determine_children(node_path, main_tree, is_leaf)
}
\arguments{
\item{node_path}{A character vector representing the path to the node.}

\item{main_tree}{Main tree structure containing the XML data.}

\item{is_leaf}{Boolean indicating if the node is a leaf.}
}
\value{
A list of children.
}
\description{
Enhances readability by encapsulating the logic for understanding the node's
offspring.
}
\keyword{internal}
