% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree.R
\name{determine_leaf_status}
\alias{determine_leaf_status}
\title{Identify if a Node is a Leaf}
\usage{
determine_leaf_status(node_path, main_tree)
}
\arguments{
\item{node_path}{A character vector representing the path to the node.}

\item{main_tree}{Main tree structure containing the XML data.}
}
\value{
Boolean indicating if the node is a leaf.
}
\description{
Essential to understand the tree structure and facilitates further
decision-making about the node.
}
\keyword{internal}
