% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree.R
\name{determine_scale_label}
\alias{determine_scale_label}
\title{Determine a Node's Scale Label}
\usage{
determine_scale_label(node_path, main_tree)
}
\arguments{
\item{node_path}{A character vector representing the path to the node.}

\item{main_tree}{Main tree structure containing the XML data.}
}
\value{
Scale label of the node.
}
\description{
Useful for annotating or understanding the node at a more descriptive level.
}
\keyword{internal}
