% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree.R
\name{determine_scale_node}
\alias{determine_scale_node}
\title{Determine the Scale of a Node}
\usage{
determine_scale_node(node_path, main_tree)
}
\arguments{
\item{node_path}{A character vector representing the path to the node.}

\item{main_tree}{Main tree structure containing the XML data.}
}
\value{
Scale of the node.
}
\description{
Used to understand the node's working scale, which might be necessary for
subsequent calculations or visualization.
}
\keyword{internal}
