% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo.R
\name{estimate_mc_time}
\alias{estimate_mc_time}
\title{Estimates execution time for Monte Carlo simulations}
\usage{
estimate_mc_time(tree, num_runs, num_test = 50)
}
\arguments{
\item{tree}{Decision tree for simulations.}

\item{num_runs}{Total number of simulations desired.}

\item{num_test}{Number of test simulations for time estimation (default is
50).}
}
\value{
No explicit return. The function prints the estimated execution time.
}
\description{
Estimates the time required to run a set number of Monte Carlo simulations
based on the time taken to run a smaller test set.
}
