% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synoptique.R
\name{evaluate_and_label}
\alias{evaluate_and_label}
\title{Evaluate Tree Node and Assign Labels}
\usage{
evaluate_and_label(df, tree, option)
}
\arguments{
\item{df}{Data frame from tree.}

\item{tree}{Tree structure object.}

\item{option}{Evaluation options for the tree.}
}
\value{
Data frame with evaluations and labels.
}
\description{
Evaluate Tree Node and Assign Labels
}
