% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option.R
\name{evaluate_scenario}
\alias{evaluate_scenario}
\title{Evaluate node values in a tree}
\usage{
evaluate_scenario(tree, option)
}
\arguments{
\item{tree}{A Tree object.}

\item{option}{Matrix representation of a scenario, providing values for the
tree's leaves.}
}
\value{
Numeric vector representing the evaluated values for all nodes, from
leaves to root.
}
\description{
Calculate the values of the tree's nodes by progressing from the leaves
(values provided by the option) up to the root. This bottom-up approach
ensures that each node's value is determined considering the values of its
child nodes.
}
