% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option.R
\name{evaluate_scenarios}
\alias{evaluate_scenarios}
\title{Evaluate multiple scenarios for a given tree}
\usage{
evaluate_scenarios(tree, options_matrix)
}
\arguments{
\item{tree}{Tree object for evaluation.}

\item{options_matrix}{Matrix where each column represents a scenario.}
}
\value{
List of numeric vectors with evaluation results for each scenario.
}
\description{
Uses \code{evaluate_scenario} to evaluate multiple scenarios simultaneously.
Each scenario is represented as a column in the \code{options_matrix}.
}
