% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oat.R
\name{evaluate_variation}
\alias{evaluate_variation}
\title{positive and negative deviations. It yields the evaluation outcome or returns
-1 if the deviation is not within the allowed range.}
\usage{
evaluate_variation(tree, option, leaf_index, variation, range_scale)
}
\arguments{
\item{tree}{A \code{Tree} object designated for assessment.}

\item{option}{Configuration set for potential alteration.}

\item{leaf_index}{\code{numeric} specifying the index of the \code{Node}
(leaf) to be varied.}

\item{variation}{\code{numeric} value detailing the magnitude and course of
the deviation (+1 or -1).}

\item{range_scale}{\code{numeric} indicating the maximum permissible scope
for the variation.}
}
\value{
A \code{numeric} value representing the assessment result of the
variation, or -1 if the deviation surpasses the set boundary.
}
\description{
positive and negative deviations. It yields the evaluation outcome or returns
-1 if the deviation is not within the allowed range.
}
\keyword{internal}
