% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree.R
\name{EvaluateOrder}
\alias{EvaluateOrder}
\title{Evaluate the order of nodes in a tree}
\usage{
EvaluateOrder(aTree)
}
\arguments{
\item{aTree}{A tree}
}
\value{
The evaluated order of nodes
}
\description{
Evaluate the order of nodes in a tree
}
