% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option.R
\name{EvaluateScenario}
\alias{EvaluateScenario}
\title{Evaluates a single scenario for a given tree}
\usage{
EvaluateScenario(aTree, option)
}
\arguments{
\item{aTree}{A "Tree" object}

\item{option}{A matrix representing a single option to evaluate}
}
\value{
A numeric vector representing the evaluation results of the scenario
}
\description{
Evaluates a single scenario by assigning values to the leaves of the tree
according to the given option and calculates the values of the aggregated
nodes.
}
