% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_subtree.R
\name{extract_names}
\alias{extract_names}
\title{Extract Node Names from Tree Nodes}
\usage{
extract_names(tree_nodes, is_leaf = TRUE)
}
\arguments{
\item{tree_nodes}{List of nodes from which names are extracted.}

\item{is_leaf}{Logical value to determine if names are extracted from leaf
nodes or non-leaf nodes.}
}
\value{
A vector of node names.
}
\description{
Retrieves names of nodes based on their leaf status.
}
