% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov.R
\name{generate_aov_formula}
\alias{generate_aov_formula}
\title{Construct AOV Formula from Results Dataframe}
\usage{
generate_aov_formula(results_df, order = 1)
}
\arguments{
\item{results_df}{\code{data.frame} containing the results for which an AOV
formula is to be constructed.}

\item{order}{\code{numeric} value specifying the interaction order: 1 for
first-order and 2 for second-order.}
}
\value{
A \code{formula} object suitable for AOV analysis.
}
\description{
Builds an Analysis of Variance (AOV) formula based on a provided results
dataframe and the desired order of interaction.
}
\keyword{internal}
