% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synoptique.R
\name{get_box_coordinates}
\alias{get_box_coordinates}
\title{Extract Box Coordinates for Plotting}
\usage{
get_box_coordinates(df)
}
\arguments{
\item{df}{Data frame with depth and size attributes.}
}
\value{
Data frame with box plotting coordinates.
}
\description{
Extract Box Coordinates for Plotting
}
