% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree.R
\name{get_chaine}
\alias{get_chaine}
\title{Retrieve Node Chain}
\usage{
get_chaine(node_path)
}
\arguments{
\item{node_path}{\code{character} vector indicating the node's path.}
}
\value{
\code{character} representing the node chain.
}
\description{
Extracts the chain of nodes based on the provided node path.
}
\keyword{internal}
