% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{get_estimated_weights}
\alias{get_estimated_weights}
\title{Get Estimated Weights for Node}
\usage{
get_estimated_weights(node)
}
\arguments{
\item{node}{A Node object.}
}
\value{
A numeric vector of estimated weights.
}
\description{
\code{get_estimated_weights} takes a \code{Node} object and calculates weights for each
of its leaves.
}
\details{
The process utilizes a linear regression model, where the root node's
aggregation table values form the response variable (denoted as 'y'), and the
leaf values are the predictors ('x').

Coefficients derived from the model, excluding the intercept, signify the
relative contributions of each leaf to the root node. Normalizing these
coefficients to sum up to 1 provides an estimate of the weight for each leaf.
}
