% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree.R
\name{get_id}
\alias{get_id}
\title{Get the ID of a given node}
\usage{
get_id(list_nodes, node_name)
}
\arguments{
\item{list_nodes}{A list of nodes}

\item{node_name}{The name of the node}
}
\value{
The ID of the node
}
\description{
Retrieves the ID of a given node from a list of nodes.
}
