% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_index.R
\name{get_node_id}
\alias{get_node_id}
\title{Retrieve Node Identifier by Name from Decision Tree}
\usage{
get_node_id(tree, node_name, avoid_repetition)
}
\arguments{
\item{tree}{A \code{Tree} object containing the decision tree structure.}

\item{node_name}{\code{character} specifying the name of the node for which
the ID is sought.}

\item{avoid_repetition}{\code{logical} indicating whether to return a single
ID in cases of node name repetition. Defaults to FALSE.}
}
\value{
A \code{numeric} representing the identifier of the specified node
within the tree.
}
\description{
Fetches the unique identifier of a specific node within the provided decision
tree based on the node's name. Takes into account potential name repetitions
and the hierarchical structure of the tree.
}
\keyword{internal}
