% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree.R
\name{get_paths}
\alias{get_paths}
\title{Retrieve Attribute Paths in an XML Tree}
\usage{
get_paths(attributes, main_tree, root_name)
}
\arguments{
\item{attributes}{A character vector containing the names of the attributes
for which to find paths.}

\item{main_tree}{An XML object representing the tree within which to
search for paths.}

\item{root_name}{A character representing the name of the root attribute
from which to begin the search.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{        A list of paths, where each path is a character vector
}\if{html}{\out{</div>}}

representing the sequence of attributes from the root name to the target
attribute.
}
\description{
Traverses an XML tree (representing a hierarchical structure) to find
specific paths for a given list of attributes, starting from a specified root
name.
}
