% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_index.R
\name{get_sensitivity_index}
\alias{get_sensitivity_index}
\title{Compute Sensitivity Index for Decision Tree}
\usage{
get_sensitivity_index(tree, avoid_repetition = FALSE)
}
\arguments{
\item{tree}{A \code{Tree} object on which the analysis is to be performed.}

\item{avoid_repetition}{\code{logical} indicating if repeated nodes should be
disregarded. Defaults to FALSE.}
}
\value{
A \code{vector} containing the Sensitivity Indices for each attribute
in the tree.
}
\description{
Computes the Sensitivity Index (SI) for each attribute within the specified
tree.
}
