% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree.R
\name{getChaine}
\alias{getChaine}
\title{Get the chain of nodes}
\usage{
getChaine(listeNoeuds)
}
\arguments{
\item{listeNoeuds}{A list of nodes}
}
\value{
A string representing the chain of nodes
}
\description{
Retrieves the chain of nodes from a given list of nodes.
}
