% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{getEstimatedWeights}
\alias{getEstimatedWeights}
\title{Title}
\usage{
getEstimatedWeights(aNode)
}
\arguments{
\item{aNode}{aNode}
}
\description{
Calculate the estimated weights from the aggregation table
}
