% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree.R
\name{getID}
\alias{getID}
\title{Get the ID of a given node}
\usage{
getID(listNodes, nodeName)
}
\arguments{
\item{listNodes}{A list of nodes}

\item{nodeName}{The name of the node}
}
\value{
The ID of the node
}
\description{
Retrieves the ID of a given node from a list of nodes.
}
