% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree.R
\name{getLeaves}
\alias{getLeaves}
\title{Get the leaves of a given node}
\usage{
getLeaves(aTree, nodeID)
}
\arguments{
\item{aTree}{A Tree}

\item{nodeID}{The ID of the node}
}
\value{
The leaves of the node
}
\description{
Retrieves the leaves of a given node
}
