% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_subtree.R
\name{handle_leaf_aggregated}
\alias{handle_leaf_aggregated}
\title{Handle Leaf-Aggregated Nodes in a Tree}
\usage{
handle_leaf_aggregated(tree, tree_nodes, leaf_names, aggregated_names)
}
\arguments{
\item{tree}{Original Tree object.}

\item{tree_nodes}{List of tree nodes under consideration.}

\item{leaf_names}{Names of leaf nodes.}

\item{aggregated_names}{Names of aggregated nodes.}
}
\value{
A vector containing names of nodes that are both leaves and
aggregated.
}
\description{
Manages nodes that are both leaves and aggregated within the tree structure.
}
