% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_subtree.R
\name{handle_multiple}
\alias{handle_multiple}
\title{Process and Handle Multiple Nodes}
\usage{
handle_multiple(tree, tree_nodes, leaf_names)
}
\arguments{
\item{tree}{Original Tree object.}

\item{tree_nodes}{List of tree nodes under consideration.}

\item{leaf_names}{Names of leaf nodes.}
}
\value{
A matrix detailing occurrences of nodes or a data frame if no
multiple nodes.
}
\description{
Manages scenarios where nodes might appear multiple times within the tree
structure.
}
