% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonteCarlo.R
\name{infoMC}
\alias{infoMC}
\title{Title}
\usage{
infoMC(aTree, nbRuns, iTest = 50)
}
\arguments{
\item{aTree}{aTree}

\item{nbRuns}{nbRuns}

\item{iTest}{iTest}
}
\description{
A simple function to give information on time requested to perform the MC
}
