% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Options.R
\name{loadOptions}
\alias{loadOptions}
\title{Load a table of options from a file}
\usage{
loadOptions(filename)
}
\arguments{
\item{filename}{The name of the file to load the options table from}
}
\value{
A matrix representing the loaded options table
}
\description{
Retrieves previously saved options from a file for further analysis or processing.
}
