% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo.R
\name{monte_carlo}
\alias{monte_carlo}
\title{Performs a Monte Carlo simulation}
\usage{
monte_carlo(tree, num_runs, write_to_file = FALSE, verbose = TRUE)
}
\arguments{
\item{tree}{Decision tree for the simulation.}

\item{num_runs}{Number of Monte Carlo simulations to conduct.}

\item{write_to_file}{Logical determining if results should be saved to a file.}

\item{verbose}{Logical determining if additional information should be printed.}
}
\value{
A matrix representing the Monte Carlo simulation results.
}
\description{
Conducts a Monte Carlo simulation on a given decision tree with a specified
number of runs. If \code{write_to_file} is set to TRUE, a .csv file named "MC
options.csv" is saved, containing the random options selected for the Monte Carlo analysis.
}
