% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonteCarlo.R
\name{MonteCarlo}
\alias{MonteCarlo}
\title{Title}
\usage{
MonteCarlo(aTree, nbRuns, isFile = F, verbose = T)
}
\arguments{
\item{aTree}{aTree}

\item{nbRuns}{nbRuns}

\item{isFile}{isFile}

\item{verbose}{verbose}
}
\description{
MonteCarlo approach
}
