% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oat.R
\name{oat}
\alias{oat}
\title{Perform One-Factor-At-A-Time (OFAT) Sensitivity Analysis}
\usage{
oat(tree, option)
}
\arguments{
\item{tree}{A decision tree object to analyze.}

\item{option}{Initial configuration for tree parameters.}
}
\value{
A matrix with evaluation results for each attribute in the tree,
under different parameter variations.
}
\description{
Conducts OFAT sensitivity analysis by performing
simulations with varying factors, keeping all but one factor constant
for each simulation.
}
