% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_subtree.R
\name{obtain_attribute_ids}
\alias{obtain_attribute_ids}
\title{Identify Descendant Attribute IDs of a Node}
\usage{
obtain_attribute_ids(tree, node_id)
}
\arguments{
\item{tree}{A Tree object.}

\item{node_id}{Node ID that serves as the reference point.}
}
\value{
A vector of attribute IDs representing descendant nodes.
}
\description{
Retrieves all nodes that share a path with the specified node. Such nodes are
considered descendants of the given node and are included in the subtree.
}
