% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_index.R
\name{plot_sensitivity_index}
\alias{plot_sensitivity_index}
\title{Show Sensitivity Index (SI)}
\usage{
plot_sensitivity_index(tree, sensitivity_indices)
}
\arguments{
\item{tree}{A decision tree object for analysis.}

\item{sensitivity_indices}{A vector of Sensitivity Indices for each leaf
in the decision tree.}
}
\value{
No return value, but generates a bar plot.
}
\description{
Generates a bar plot to visualize the Sensitivity Index (SI) of the leaves of
a given decision tree.
}
