% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{print,Node-method}
\alias{print,Node-method}
\alias{print.Node}
\alias{\S4method{print}{Node}}
\title{Print Method for Node Class Object}
\usage{
\S4method{print}{Node}(x, ...)
}
\arguments{
\item{x}{An object of class \code{Node} that you want to print.}

\item{...}{Additional arguments to be passed to the underlying print
function, though they might not have any effect in this custom print
method.}
}
\value{
This function is invoked for its side effect of printing. It does not
return anything.
}
\description{
Provides a comprehensive display of a Node's properties, such as its name,
ID, depth, path, and more. This method is intended for better readability
and understanding of a Node's structure and relationships.
}
\seealso{
\itemize{
\item \code{\link{Node-class}}: For more details on the Node class.
}
}
