% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{print,Tree-method}
\alias{print,Tree-method}
\alias{print.Tree}
\title{print.tree}
\usage{
\S4method{print}{Tree}(x, ...)
}
\arguments{
\item{x}{The printed Tree}

\item{...}{see '...' print parameters}
}
\description{
Print method for Tree class object
}
