% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree.R
\name{process_aggregated_leaf}
\alias{process_aggregated_leaf}
\title{Process Leaf and Aggregated Nodes}
\usage{
process_aggregated_leaf(tree_nodes, leaves, aggregated)
}
\arguments{
\item{tree_nodes}{A list of tree nodes.}

\item{leaves}{A vector containing the names of leaf nodes.}

\item{aggregated}{A vector containing the names of aggregated nodes.}
}
\value{
A list with the following elements:
\itemize{
\item \code{leaf_aggregated}: A vector containing the names of nodes
that are both leaf and aggregated.
\item \code{is_leaf_aggregated}: A logical value indicating if such nodes
exist.
\item \code{tree_nodes}: A list of tree nodes updated with the twin
information and markings for leaf and aggregated nodes.
}
}
\description{
Identifies and processes the nodes in the tree that are both leaf and
aggregated. Such nodes can have specific roles in decision or aggregation
trees and may require special treatment. By identifying these nodes, the
function aids in properly handling them in the overall evaluation or
traversal of the tree.
}
