% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree.R
\name{process_duplicated_leaf}
\alias{process_duplicated_leaf}
\title{Process Duplicate Leaf Nodes in Tree Structure}
\usage{
process_duplicated_leaf(tree_nodes, leaves)
}
\arguments{
\item{tree_nodes}{A \code{list} of nodes, where each element is an object of
class \code{Node}, representing the tree's hierarchical structure.}

\item{leaves}{A \code{character} vector detailing the names of leaf
attributes within the tree.}
}
\value{
A list containing the following components:
\itemize{
\item \code{multiple}: A \code{data.frame} or \code{matrix} presenting the
count of occurrences for leaf attributes with more than one
appearance.
\item \code{is_multiple}: A \code{logical} indicating the presence of leaf
attributes with multiple occurrences.
\item \code{tree_nodes}: An updated \code{list} of tree nodes, where each
element is an object of class \code{Node}, enriched with
\code{Twin} information for duplicated leaf attributes.
}
}
\description{
An internal utility to identify multiple occurrences of leaf
attributes within a tree and mark them with the appropriate \code{Twin}
reference for differentiation. It also provides a comprehensive overview of
these occurrences, enhancing the process of tree traversal and analysis when
invoked by main functions like \code{\link{create_tree}}.
}
\seealso{
\code{\link{create_tree}} for the primary function that utilizes
this utility in tree construction.
}
\keyword{internal}
