% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option.R
\name{save_options}
\alias{save_options}
\title{Save options table}
\usage{
save_options(options_table, file_name)
}
\arguments{
\item{options_table}{Matrix containing option values.}

\item{file_name}{Desired file name for saving the options.}
}
\description{
Stores a matrix of options into a file, primarily for archival or subsequent
analysis.
}
