% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Options.R
\name{saveOptions}
\alias{saveOptions}
\title{Save a table of options to a file}
\usage{
saveOptions(anOptionsTable, aFileName)
}
\arguments{
\item{anOptionsTable}{A matrix representing options}

\item{aFileName}{The name of the file to save the options table}
}
\description{
Stores a table of options into a file for later use or analysis.
}
