% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option.R
\name{saveScenarios}
\alias{saveScenarios}
\title{Save evaluation results of scenarios to a file}
\usage{
saveScenarios(theScenarios, file)
}
\arguments{
\item{theScenarios}{A list of numeric vectors representing the evaluation
results of the scenarios}

\item{file}{The name of the file to save the scenarios}
}
\description{
Stores the results of scenario evaluations into a file for later analysis.
}
