% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov.R
\name{sensib.grpe}
\alias{sensib.grpe}
\title{Calculate sensitivity factors for model terms}
\usage{
sensib.grpe(aov.obj, fact.interet, las = 1, ...)
}
\arguments{
\item{aov.obj}{An object of class \code{aov} resulting from a call to \code{aov()}.}

\item{fact.interet}{Factors of interest.}

\item{las}{The style of axis labels on the bar plot. Default is 1.}

\item{...}{Additional arguments to control the appearance of the bar plot.}
}
\value{
A data frame with sensitivity factors for each term in the model.
}
\description{
Calculates sensitivity factors for each term in a fitted model based on
specific factors of interest. The function uses the parameters \code{las} and
\code{...} to control the appearance of the generated bar plot.
}
