% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov.R
\name{sensib.total}
\alias{sensib.total}
\title{Calculate sensitivity factors for model terms}
\usage{
sensib.total(aov.obj)
}
\arguments{
\item{aov.obj}{An object of class \code{aov} resulting from a call to \code{aov()}.}
}
\value{
A data frame with sensitivity factors for each term in the model.
}
\description{
Calculates sensitivity factors for each term in a fitted model.
}
