% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{show,Tree-method}
\alias{show,Tree-method}
\alias{show.Tree}
\title{show method for Tree class object}
\usage{
\S4method{show}{Tree}(object)
}
\arguments{
\item{object}{The Tree object to be shown.}
}
\description{
Custom show method for Tree class object. Prints a formatted structure of the
tree using a set of rules based on node properties.
}
