% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo.R
\name{show_mc_results}
\alias{show_mc_results}
\title{Displays the Monte Carlo simulation results}
\usage{
show_mc_results(node, mc_results, num_runs)
}
\arguments{
\item{node}{Node of interest.}

\item{mc_results}{Monte Carlo simulation results matrix.}

\item{num_runs}{Number of Monte Carlo simulations conducted.}
}
\value{
A vector representing the bar chart data.
}
\description{
Visualizes the Monte Carlo simulation outcomes for a given node as a bar chart,
depicting the frequency of each outcome. Additionally, the bar lengths are saved
to a .csv file named "MC bar lengths.csv".
}
