% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oat.R
\name{show_oat_results}
\alias{show_oat_results}
\title{Visualize One-Factor-At-A-Time (OFAT) Sensitivity Analysis Results}
\usage{
show_oat_results(node_name, results, tree)
}
\arguments{
\item{node_name}{Name of the node for the visualization.}

\item{results}{Evaluation results matrix obtained from \code{ofat_sensitivity_analysis}.}

\item{tree}{Decision tree object used in the OFAT analysis.}
}
\value{
No return value; the function displays a plot.
}
\description{
Generates a plot visualizing the results of the OFAT sensitivity analysis.
}
