% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option.R
\name{show_scenario}
\alias{show_scenario}
\title{Plot a bar chart for a single scenario}
\usage{
show_scenario(scenario, tree, label_y = TRUE, modify_par = TRUE)
}
\arguments{
\item{scenario}{Scenario data to visualize.}

\item{tree}{Associated Tree object providing attribute details.}

\item{label_y}{Logical value indicating whether to label the Y-axis (default is TRUE).}

\item{modify_par}{Logical value to decide if graphical parameters should be modified (default is TRUE).}
}
\description{
Visualizes the attribute values of a provided scenario. For each attribute,
a bar is plotted, and the maximum possible value is highlighted.
}
\seealso{
\code{\link{evaluate_scenario}}
}
