% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AOV.R
\name{showAOV}
\alias{showAOV}
\title{Title}
\usage{
showAOV(aAOV_DEXi, main.show = T, nb.plot = 8, beside = T, las = 1, ...)
}
\arguments{
\item{aAOV_DEXi}{aAOV_DEXi}

\item{main.show}{main.show}

\item{nb.plot}{nb.plot}

\item{beside}{beside}

\item{las}{las}

\item{...}{...}
}
\description{
Title
}
