% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo.R
\name{ShowMC}
\alias{ShowMC}
\title{Displays the Monte Carlo simulation results}
\usage{
ShowMC(Node, MC, nbRuns)
}
\arguments{
\item{Node}{Node of interest}

\item{MC}{Monte Carlo simulation results}

\item{nbRuns}{Number of Monte Carlo simulations performed}
}
\value{
Returns the bar chart data
}
\description{
This function generates a bar chart of the frequency of each outcome in the
Monte Carlo simulation results for a given node. It also writes the bar
lengths to a .csv file.
}
