% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OAT.R
\name{showOAT}
\alias{showOAT}
\title{Visualize One-Factor-At-A-Time (OFAT) Sensitivity Analysis Results}
\usage{
showOAT(nodeName, aResults, aTree)
}
\arguments{
\item{nodeName}{Name of the node for which to generate the plot.}

\item{aResults}{Matrix of evaluation results returned by the OAT function.}

\item{aTree}{A decision tree object that was used in the OAT function.}
}
\value{
No return value, but generates a plot.
}
\description{
Generates a plot to visualize the results of One-Factor-At-A-Time (OFAT) sensitivity analysis.
}
