% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Options.R
\name{showScenario}
\alias{showScenario}
\title{Title}
\usage{
showScenario(aScenario, aTree, isLabelY = TRUE, isPar = T)
}
\arguments{
\item{aScenario}{the scenario to graph}

\item{aTree}{the tree}

\item{isLabelY}{do we want the label of the option}

\item{isPar}{do we want the function to modify the par}
}
\description{
A function to create a graph of a scenario as horizontal bar
}
