% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_index.R
\name{si_dexi}
\alias{si_dexi}
\title{Calculate Sensitivity Index (SI)}
\usage{
si_dexi(
  tree,
  file_name = "SI_out.csv",
  is_file = TRUE,
  avoid_repetition = FALSE
)
}
\arguments{
\item{tree}{A decision tree object to perform the analysis on.}

\item{file_name}{The file name to write the SI to. Default is "SI_out.csv".}

\item{is_file}{A boolean to decide whether to write the SI to a file. Default
is TRUE.}

\item{avoid_repetition}{A boolean to decide whether to avoid repeated nodes.
Default is FALSE.}
}
\value{
A list of Sensitivity Indices for each node in the tree.
}
\description{
Calculates the Sensitivity Index (SI) for a given decision tree.
}
