% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov.R
\name{visualize_aov_results}
\alias{visualize_aov_results}
\title{Visualize AOV results}
\usage{
visualize_aov_results(
  aov_results,
  show_main = TRUE,
  num_plots = 8,
  horizontal = TRUE,
  axis_label_style = 1,
  ...
)
}
\arguments{
\item{aov_results}{A list containing the AOV results.}

\item{show_main}{Logical, if TRUE, main effects and total sums of squares are
displayed in the plot. Defaults to TRUE.}

\item{num_plots}{The number of effects to display.}

\item{horizontal}{Logical, if TRUE, the bar plot is displayed horizontally.
Defaults to TRUE.}

\item{axis_label_style}{The style of axis labels on the bar plots. Default is
1.}

\item{...}{Additional arguments to control the appearance of the bar plots.}
}
\value{
A data frame with the summed square proportions.
}
\description{
Visualizes the results of an Analysis of Variance (AOV). It provides both a
complete total and effect visualization using bar plots.
}
