% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_subtree.R
\name{adjust_for_leaf_aggregated}
\alias{adjust_for_leaf_aggregated}
\title{Adjust Node ID for Trees with Leaf-Aggregated Structures}
\usage{
adjust_for_leaf_aggregated(tree, node_id)
}
\arguments{
\item{tree}{a \code{Tree} object that might possess a leaf-aggregated
structure.}

\item{node_id}{a \code{numeric} or \code{character} representing the
identifier for the \code{Node} to be adjusted.}
}
\value{
A \code{numeric} vector encompassing the adjusted node IDs.
}
\description{
When a \code{Tree} has a leaf-aggregated structure, the node ID may require
adjustment. This function carries out the necessary transformation for such
nodes.
}
\keyword{internal}
