% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_index.R
\name{adjust_probas_for_range_scale}
\alias{adjust_probas_for_range_scale}
\title{Adjust Probabilities for Range Scale}
\usage{
adjust_probas_for_range_scale(probs, node)
}
\arguments{
\item{probs}{The computed probabilities.}

\item{node}{The specific node for which the adjustment should be made.}
}
\value{
A list of adjusted probabilities.
}
\description{
Adjusts the computed probabilities to match the range scale of a node.
}
