% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov.R
\name{aov_tree}
\alias{aov_tree}
\title{Dual Order AOV on a Decision Tree}
\usage{
aov_tree(tree)
}
\arguments{
\item{tree}{\code{Tree} object to be analyzed.}
}
\value{
A \code{list} containing results for both first-order and
second-order AOV analyses.
}
\description{
Conducts an Analysis of Variance (AOV) on a provided decision tree, computing
both first-order and second-order effects.
}
\examples{
tree <- dexiranalysis::masc2
subtree <- create_sub_tree(tree, "Dimension sociale")
AOV_out <- aov_tree(subtree)

}
