% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree.R
\name{are_leaves_unevaluated}
\alias{are_leaves_unevaluated}
\title{Check if Any Leaves are Unevaluated}
\usage{
are_leaves_unevaluated(leaves, results, nodes)
}
\arguments{
\item{leaves}{A vector of leaves to check.}

\item{results}{A named vector where names are node names and values indicate
whether the node has been evaluated.}

\item{nodes}{A list of nodes containing node details.}
}
\value{
Logical. Returns \code{TRUE} if any leaves are unevaluated, and \code{FALSE}
otherwise.
}
\description{
Determines if any of the provided leaves are marked as unevaluated in the
results. A leaf is considered unevaluated if its corresponding value in the
results is -1.
}
\keyword{internal}
