% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option.R
\name{assign_values_to_leaves}
\alias{assign_values_to_leaves}
\title{Assign values to the leaves of the tree}
\usage{
assign_values_to_leaves(results, option)
}
\arguments{
\item{results}{Numeric vector placeholder for results.}

\item{option}{A matrix representing a single option/scenario to evaluate.}
}
\value{
Numeric vector with leaf-populated values.
}
\description{
Ensure that there are no Leaf-Aggregated leaves in the given option.
}
