% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_index.R
\name{calculate_conditional_probabilities}
\alias{calculate_conditional_probabilities}
\title{Calculate Conditional Probabilities}
\usage{
calculate_conditional_probabilities(input_table, weight_list, sy, y_levels)
}
\arguments{
\item{input_table}{A matrix with all the level combinations of the A_i factors
and, in the last column, the associated Y values.}

\item{weight_list}{A list with the weight vectors of each A_i variable.
If missing, A_i levels have equal weights.}

\item{sy}{The number of unique Y values.}

\item{y_levels}{Optional, specifies the Y levels. If missing, they are
extracted from the input_table.}
}
\value{
A list containing matrices of Y probabilities conditional to each
A_i factor and, finally, the vector of marginal Y probabilities.
}
\description{
Calculates the probabilities of Y given its direct descendants A_i.
}
