% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_index.R
\name{calculate_sensitivity_indices}
\alias{calculate_sensitivity_indices}
\title{Calculate Sensitivity Indices (SI) From Conditional Probabilities}
\usage{
calculate_sensitivity_indices(conditional_prob_list, weight_list)
}
\arguments{
\item{conditional_prob_list}{A list containing matrices of conditional
probabilities (of \( Y \) conditional on each \( A_i \)) and a vector of \(
Y \) probabilities as the last element.}

\item{weight_list}{A list representing the weights associated with each \(
A_i \) factor level.}
}
\value{
A named vector of Sensitivity Indices (SI) where names correspond to
the factors.
}
\description{
Computes the first-order sensitivity indices of \( Y \) with respect to \(
A_i \) descendants, utilizing the provided conditional probabilities and \(
A_i \) weights.
}
\keyword{internal}
