% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_index.R
\name{clcSI_DEXi}
\alias{clcSI_DEXi}
\title{Calculate Sensitivity Index (SI) Helper Function}
\usage{
clcSI_DEXi(aTree, avoidrep = F)
}
\arguments{
\item{aTree}{A decision tree object to perform the analysis on.}

\item{avoidrep}{A boolean to decide whether to avoid repeated nodes. Default
is FALSE.}
}
\value{
A vector of Sensitivity Indices for each attribute in the decision
tree.
}
\description{
A helper function for \code{SI_DEXi} to calculate Sensitivity Index (SI) of a
decision tree.
}
