% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option.R
\name{compareScenario}
\alias{compareScenario}
\title{Compare scenarios using a radial plot}
\usage{
compareScenario(aTree, theScenarios, listNodes)
}
\arguments{
\item{aTree}{A "Tree" object}

\item{theScenarios}{A list of numeric vectors representing the evaluation
results of the scenarios}

\item{listNodes}{A list of node names to include in the comparison}
}
\description{
Visualizes the comparison of values assigned to a list of nodes across
multiple scenarios using a radial plot.
}
\seealso{
\code{\link{evaluate_scenarios}}
}
