% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option.R
\name{compute_aggregated_values}
\alias{compute_aggregated_values}
\title{Calculate Values for Leaf-Aggregated Scenarios}
\usage{
compute_aggregated_values(tree, results)
}
\arguments{
\item{tree}{A Tree object.}

\item{results}{Numeric vector placeholder for results.}
}
\value{
Numeric vector with aggregated values.
}
\description{
Considers tree's evaluation order to derive aggregated values.
}
