% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov.R
\name{compute_aov_sensitivity_effects}
\alias{compute_aov_sensitivity_effects}
\title{Calculate sensitivity criteria for model terms}
\usage{
compute_aov_sensitivity_effects(aov_obj)
}
\arguments{
\item{aov_obj}{An object of class \code{aov} from a call to \code{aov()}.}
}
\value{
A data frame with sensitivity criteria for each term in the model.
The columns of the data frame include:
\itemize{
\item \code{df}: degrees of freedom
\item \code{ss}: sum of squares
\item \code{ss.ratio}: ratio of sum of squares to total sum of squares
\item \code{cm}: mean square (or variance) for each term
\item \code{F}: F-statistic for each term
}

Rows are ordered in decreasing order of ss.ratio.
}
\description{
Provides sensitivity criteria for each term in a fitted model.
}
