% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov.R
\name{compute_aov_total_sensitivity}
\alias{compute_aov_total_sensitivity}
\title{Compute Total Sensitivity Factors for Model Terms}
\usage{
compute_aov_total_sensitivity(aov_obj)
}
\arguments{
\item{aov_obj}{An \code{aov} object resulting from a call to the \code{aov()}
function.}
}
\value{
A \code{data.frame} detailing the total sensitivity factors for each
model term.
}
\description{
Derives the total sensitivity factors for each term present in a provided
fitted model.
}
\keyword{internal}
