% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_index.R
\name{compute_direct_probabilities}
\alias{compute_direct_probabilities}
\title{Compute Direct Probabilities}
\usage{
compute_direct_probabilities(node, child_weights)
}
\arguments{
\item{node}{The specific node for which the direct probabilities should be
calculated.}

\item{child_weights}{The weights of the child nodes.}
}
\value{
A list of computed direct probabilities.
}
\description{
Computes the direct probabilities for children of a node based on its
aggregation and child weights.
}
