% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option.R
\name{compute_final_aggregated_values}
\alias{compute_final_aggregated_values}
\title{Compute final aggregated values}
\usage{
compute_final_aggregated_values(tree, results)
}
\arguments{
\item{tree}{Tree object.}

\item{results}{Numeric vector placeholder for results.}
}
\value{
Numeric vector with final values.
}
\description{
Uses tree structure to finalize aggregation.
}
