% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_index.R
\name{compute_indirect_probabilities}
\alias{compute_indirect_probabilities}
\title{Compute Indirect Probabilities}
\usage{
compute_indirect_probabilities(
  node,
  probs,
  conditional_probabilities,
  tree,
  avoid_repetition
)
}
\arguments{
\item{node}{The \code{Node} object for which the indirect probabilities are
calculated.}

\item{probs}{A \code{list} of computed probabilities for the node.}

\item{conditional_probabilities}{A \code{list} of existing conditional
probabilities associated with the node's descendants.}

\item{tree}{The \code{Tree} object representing the decision tree structure.}

\item{avoid_repetition}{A \code{logical} flag determining if nodes with
repeated names should be avoided in the calculations.}
}
\value{
A \code{list} of computed indirect probabilities for the node's
descendants.
}
\description{
Computes the indirect probabilities for descendants of a specific node,
considering the computed probabilities of the node, the existing conditional
probabilities, and the structure of the decision tree. If a child of the node
is not a leaf, its indirect probabilities are calculated based on its direct
ancestors' probabilities.
}
\keyword{internal}
