% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option.R
\name{compute_values_from_aggregation_table}
\alias{compute_values_from_aggregation_table}
\title{Compute Node Values from Aggregation Table}
\usage{
compute_values_from_aggregation_table(tree, results)
}
\arguments{
\item{tree}{A \code{Tree} object.}

\item{results}{A \code{numeric} vector with pre-existing values.}
}
\value{
A \code{numeric} vector with updated node values.
}
\description{
Calculates node values iteratively using the aggregation table and related children.
}
\keyword{internal}
