% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{create_aggregation_matrix}
\alias{create_aggregation_matrix}
\title{Aggregation Matrix Creation via Genetic Algorithm}
\usage{
create_aggregation_matrix(
  node,
  expected_weights,
  number_of_tables = 1,
  population_size = 50,
  iterations = 50
)
}
\arguments{
\item{node}{An object of class \code{Node}. Represents the target node.}

\item{expected_weights}{\code{numeric} vector, denoting the expected weights.}

\item{number_of_tables}{\code{numeric}. Defines the number of tables; default
is set to 1.}

\item{population_size}{\code{numeric}. Sets the population size for the
genetic algorithm; default is 50.}

\item{iterations}{\code{numeric}. Specifies the number of iterations for the
genetic algorithm; default is 50.}
}
\value{
List of aggregation matrices.
}
\description{
Optimizes an aggregation matrix for a specified node.
}
\details{
Leverages a genetic algorithm where each "gene" mirrors the value of the
aggregated node. This optimization seeks to minimize the squared discrepancy
between expected and actual weights, ensuring they align closely with
anticipated values.
}
\seealso{
\itemize{
\item \code{\link{Node-class}}: For a detailed understanding of the Node class.
\item \code{\link{rbga}}: The genetic algorithm function from the genalg package.
\item \code{\link{print.Node}}: To print a Node object.
}
}
\keyword{internal}
