% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov.R
\name{create_aov_barplot}
\alias{create_aov_barplot}
\title{Create AOV Bar Plot}
\usage{
create_aov_barplot(
  data,
  is_effect,
  horizontal,
  num_plots,
  show_main,
  axis_label_style,
  ...
)
}
\arguments{
\item{data}{A data frame containing the AOV results to be plotted.}

\item{is_effect}{A logical flag. If TRUE, indicates that the data represents
effects; otherwise, it represents totals.}

\item{horizontal}{A logical flag. If TRUE, the bar plots will be oriented
horizontally; otherwise, they will be vertical.}

\item{num_plots}{An integer specifying the number of effects to display in
the bar plot.}

\item{show_main}{A logical flag. If TRUE, the main effects and total sums of
squares are combined and displayed. This is effective only if \code{is_effect}
is FALSE.}

\item{axis_label_style}{An integer representing the style of the axis labels
on the bar plots.}

\item{...}{Additional arguments that will be passed to the \code{barplot} function
to control its appearance.}
}
\description{
Helper function specifically designed to create bar plots to visualize the
results of an AOV (Analysis of Variance). It has the flexibility to handle
both total sums of squares and effect visualizations.
}
