% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synoptique.R
\name{create_data_frame_for_plotting}
\alias{create_data_frame_for_plotting}
\title{Produce Dataframe Suitable for Plotting from Tree Structure}
\usage{
create_data_frame_for_plotting(tree)
}
\arguments{
\item{tree}{Tree structure object.}
}
\value{
Data frame suitable for plotting.
}
\description{
Produce Dataframe Suitable for Plotting from Tree Structure
}
