% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov.R
\name{create_factorial_plan}
\alias{create_factorial_plan}
\title{Create Factorial Plan from Tree}
\usage{
create_factorial_plan(tree)
}
\arguments{
\item{tree}{Tree object for analysis.}
}
\value{
A matrix representing the factorial plan.
}
\description{
Extracts leaf indices from the tree and creates the factorial plan based on
gen.factorial function.
}
