% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synoptique.R
\name{create_list_synoptique}
\alias{create_list_synoptique}
\title{Title}
\usage{
create_list_synoptique(Tree, options, depth = NA)
}
\arguments{
\item{Tree}{Tree}

\item{options}{options}

\item{depth}{depth}
}
\description{
Title
}
