% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synoptique.R
\name{create_plot}
\alias{create_plot}
\title{Create Synoptic Plot from Data Frame}
\usage{
create_plot(df2)
}
\arguments{
\item{df2}{Data frame with plotting details.}
}
\value{
ggplot object.
}
\description{
Create Synoptic Plot from Data Frame
}
