% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_subtree.R
\name{create_sub_tree}
\alias{create_sub_tree}
\title{Create a Subtree Starting from a Specific Node}
\usage{
create_sub_tree(tree, node_name, avoid_repetition = FALSE)
}
\arguments{
\item{tree}{Tree object to derive the subtree from.}

\item{node_name}{Starting node for the subtree.}

\item{avoid_repetition}{Prevent repeated nodes? Default is FALSE.}
}
\value{
Subtree as a Tree object.
}
\description{
Constructs a subtree from a specified node within a tree. The subtree
encapsulates all descendant attributes of the node.
}
