% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synoptique.R
\name{create_synoptique}
\alias{create_synoptique}
\title{Create a Synoptic Plot from a Given Tree Structure}
\usage{
create_synoptique(tree, option, depth = NA)
}
\arguments{
\item{tree}{Tree structure object.}

\item{option}{Evaluation options for the tree.}

\item{depth}{Depth of the tree; if provided, produces a sub-tree.}
}
\value{
A ggplot object representing the synoptic plot.
}
\description{
Create a Synoptic Plot from a Given Tree Structure
}
