% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synoptique.R
\name{create_synoptique}
\alias{create_synoptique}
\title{Title}
\usage{
create_synoptique(Tree, option, depth = NA)
}
\arguments{
\item{Tree}{Tree}

\item{option}{option}

\item{depth}{depth}
}
\description{
Title
}
