% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree.R
\name{create_tree}
\alias{create_tree}
\title{Tree Construction from DEXi's XML Output}
\usage{
create_tree(main_tree)
}
\arguments{
\item{main_tree}{A character or an object of class \code{XML}. Either the
path to the .dxi file containing the tree (character) or the XML object
that represents the primary tree structure and should encompass the DEXi
decision model.}
}
\value{
A list of \code{Tree} objects, each corresponding to a unique root
attribute in the DEXi XML layout. Each tree details the root attribute,
nodes, leaves, depth, and other associated data regarding the decision
model's configuration.
}
\description{
Constructs a \code{Tree} objects using the XML representation from a DEXi
model.
}
\details{
Iterates over every root attribute in the provided XML to create tree
structures that depict the hierarchical decision model.
}
