% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{createAggregationMatrix}
\alias{createAggregationMatrix}
\title{Title}
\usage{
createAggregationMatrix(
  aNode,
  expectedWeight,
  nbTables = 1,
  popSize = 50,
  iters = 50
)
}
\arguments{
\item{aNode}{aNode}

\item{expectedWeight}{expectedWeight}

\item{nbTables}{nbTables}

\item{popSize}{popSize}

\item{iters}{iters}
}
\description{
Title
}
