% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Options.R
\name{createOptions}
\alias{createOptions}
\title{Title}
\usage{
createOptions(aTree, nbOptions = 1, aSeed = -1)
}
\arguments{
\item{aTree}{A Tree}

\item{nbOptions}{Number option to create}

\item{aSeed}{seed for random number generation}
}
\description{
Title
}
