% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_subtree.R
\name{createSubTree}
\alias{createSubTree}
\title{Constructs a subtree from a given node in a tree}
\usage{
createSubTree(aTree, nodeName, avoidrep = F)
}
\arguments{
\item{aTree}{A "Tree" object}

\item{nodeName}{The name of the node from which to start constructing the
subtree}

\item{avoidrep}{A logical value indicating whether to avoid repeated nodes
(Default: FALSE)}
}
\value{
A "Tree" object representing the subtree
}
\description{
Constructs a subtree starting from a specific node in a provided tree
}
