% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth_order.R
\name{depth_order}
\alias{depth_order}
\title{Get Aggregated Attributes Ordered by Depth}
\usage{
depth_order(tree)
}
\arguments{
\item{tree}{A decision tree object.}
}
\value{
A vector of aggregated attribute names, ordered by depth.
}
\description{
Returns aggregated attributes ordered by depth. Typically
used for analysis strategy (AS).
For trees using aggregated attributes as leaves somewhere else,
the order of attributes might be modified in order to prevent mistakes or
missing values in future analyses: aggregated attributes are meant to be
assessed as soon as they can be.
}
\examples{
tree <- dexisensitivity::dexifruits_v1
dexisensitivity:::depth_order(tree)


}
\keyword{internal}
