% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth_order.R
\name{depth_order}
\alias{depth_order}
\title{Get Aggregated Attributes Ordered by Depth}
\usage{
depth_order(aTree)
}
\arguments{
\item{aTree}{A decision tree object.}
}
\value{
A vector of aggregated attribute names, ordered by depth.
}
\description{
Return aggregated attributes ordered by depth. This function is typically
used for analysis strategy (AS).
}
