% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{describe,Tree-method}
\alias{describe,Tree-method}
\alias{describe.Tree}
\title{Describe Method for Tree Class Objects}
\usage{
\S4method{describe}{Tree}(object)
}
\arguments{
\item{object}{The Tree object to be described.}
}
\value{
This function is primarily executed for its side effect of presenting
nodes from the Tree object and does not provide a meaningful return value.
}
\description{
Outputs a detailed structure of a Tree object, with each node and its
properties displayed distinctly.
}
\details{
Each node within the Tree is presented separately. If the Tree lacks
nodes, an error "Tree without any node!" is raised.
}
\seealso{
\itemize{
\item \code{\link{Tree-class}}: For the Tree class definition.
\item \code{\link{print.Tree}}: For printing a Tree object.
\item \code{\link{show.Tree}}: For showing a Tree object.
}
}
