% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{describe,Tree-method}
\alias{describe,Tree-method}
\alias{describe.Tree}
\title{describe method for Tree class object}
\usage{
\S4method{describe}{Tree}(object)
}
\arguments{
\item{object}{Object}
}
\value{
Primarily invoked for its side effect of displaying the nodes of the
Tree object, the function does not produce a meaningful return value.
}
\description{
A method to describe the structure of a Tree object in a comprehensive way.
It prints each node and its properties separately.
}
