% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree.R
\name{determine_aggregation}
\alias{determine_aggregation}
\title{Determine Aggregation for a Node}
\usage{
determine_aggregation(node_path, main_tree, is_leaf, children)
}
\arguments{
\item{node_path}{A character vector representing the path to the node.}

\item{main_tree}{Main tree structure containing the XML data.}

\item{is_leaf}{Boolean indicating if the node is a leaf.}

\item{children}{List of children nodes.}
}
\value{
If the node is not a leaf, the function returns the aggregation
structure, including a matrix representing the factorial plan and
transformed low function values. If the node is a leaf, the function
returns a matrix with zero.
}
\description{
crucial for understanding how information is combined at the node level. It
contains the logic for aggregation based on whether the node is a leaf and
other properties. The function retrieves the low function values from the XML
and adjusts them to numeric format. Then, it determines the scales from nodes
n-1 and creates the factorial plan or sequence depending on the number of
children. If the node is a leaf, it returns a matrix with zero.
}
\keyword{internal}
