% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree.R
\name{determine_mother}
\alias{determine_mother}
\title{Identify the Mother of a Node}
\usage{
determine_mother(node_path)
}
\arguments{
\item{node_path}{A character vector representing the path to the node.}
}
\value{
Mother of the node if exists.
}
\description{
Contributes to understanding the relationship between nodes and navigation
within the tree's structure.
}
\keyword{internal}
