% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree.R
\name{determine_scale_label}
\alias{determine_scale_label}
\title{Determine a Node's Scale Label}
\usage{
determine_scale_label(node_path, main_tree)
}
\arguments{
\item{\code{node_path}}{A \code{character} vector representing the path to
the node.}

\item{\code{main_tree}}{An XML object representing the main tree structure.}
}
\value{
A \code{character} vector indicating the scale label of the node. If
no scale label is found, the function returns an empty character vector.
}
\description{
Retrieves the descriptive scale label for a specific node in the decision
tree. This label can offer a more detailed understanding of the node's
dimension and is valuable for annotations and enhanced comprehension of the
node.
}
\keyword{internal}
