% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree.R
\name{determine_scale_node}
\alias{determine_scale_node}
\title{Determine the Scale of a Node}
\usage{
determine_scale_node(node_path, main_tree)
}
\arguments{
\item{\code{node_path}}{A \code{character} vector representing the path to
the node.}

\item{\code{main_tree}}{An XML object representing the main tree structure.}
}
\value{
An \code{integer} indicating the scale of the node. If no scale is
found, the function returns 0.
}
\description{
Retrieves the scale associated with a given node within the decision tree.
The scale provides valuable insights into the node's dimension and can be
crucial for subsequent computations and visual representations.
}
\keyword{internal}
