% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree.R
\name{determine_sisters}
\alias{determine_sisters}
\title{Find the Sisters of a Node}
\usage{
determine_sisters(node_path, main_tree)
}
\arguments{
\item{node_path}{A character vector representing the path to the node.}

\item{main_tree}{Main tree structure containing the XML data.}
}
\value{
A list of sisters.
}
\description{
Aids in understanding horizontal relationships within the tree, crucial for
some analyses.
}
\keyword{internal}
