% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree.R
\name{determine_weight_list}
\alias{determine_weight_list}
\title{Determine Weight List for a Node}
\usage{
determine_weight_list(node_path, main_tree, is_leaf, aggregation, children)
}
\arguments{
\item{node_path}{A character vector representing the path to the node.}

\item{main_tree}{Main tree structure containing the XML data.}

\item{is_leaf}{Boolean indicating if the node is a leaf.}

\item{aggregation}{Aggregation structure, possibly used in the calculation.}

\item{children}{List of children nodes.}
}
\value{
A numeric vector containing the weights for the node.
}
\description{
Calculates the weights associated with a particular node based on
various conditions, such as whether the node is a leaf and the
number of children. Weights may represent probabilities or importance
factors in the overall tree structure.
}
\keyword{internal}
