% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dexifruits_v1}
\alias{dexifruits_v1}
\title{DEXiFruits_v1 model}
\format{
An object of class \code{Tree} of length 1.
}
\source{
\url{https://means.inrae.fr/outils-emc/dexifruits/telecharger-dexifruits}
}
\usage{
dexifruits_v1
}
\description{
A Tree class object created by dexiranalysis::create_tree function, using the
Multi-Attribute Assessment model of the sustainability of cropping systems in
arboriculture (DEXiFruits) in its first and generic version.
}
\seealso{
\code{\link{Tree-class}}
}
\keyword{datasets}
