% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexiranalysis_package.R
\docType{package}
\name{dexiranalysis-package}
\alias{dexiranalysis}
\alias{dexiranalysis-package}
\title{dexiranalysis: DEXi model analysis}
\description{
Provides class and function to analyse DEXi models from DEXi software (Bohanec Marko, 1999). Tree in .dxi files can be read to create a Tree class object. This object can be vizualise and used for scenario creation, scenario evaluation and a model sensitivity analysis can be perform.
}
\author{
\strong{Maintainer}: Roland Allart \email{roland.allart@inrae.fr}

Authors:
\itemize{
  \item Jacques-Eric Bergez \email{jacques-eric.bergez@inrae.fr} (\href{https://orcid.org/0000-0003-3467-2617}{ORCID})
}

}
\keyword{internal}
