% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexisensitivity-package.R
\docType{package}
\name{dexisensitivity-package}
\alias{dexisensitivity}
\alias{dexisensitivity-package}
\title{dexisensitivity: 'DEXi' Decision Tree Analysis and Visualization}
\description{
Provides a versatile toolkit for analyzing and visualizing 'DEXi' (Decision EXpert for education) decision trees, facilitating multi-criteria decision analysis directly within R. Users can read .dxi files, manipulate decision trees, and evaluate various scenarios. It supports sensitivity analysis through Monte Carlo simulations, one-at-a-time approaches, and variance-based methods, helping to discern the impact of input variations. Additionally, it includes functionalities for generating sampling plans and an array of visualization options for decision trees and analysis results. A distinctive feature is the synoptic table plot, aiding in the efficient comparison of scenarios. Whether for in-depth decision modeling or sensitivity analysis, this package stands as a comprehensive solution. Definition of sensitivity analyses available in Carpani, Bergez and Monod (2012) \doi{10.1016/j.envsoft.2011.10.002} and detailed description of the package soon available in Alaphilippe, Allart, Carpani, Cavan, Monod and Bergez (submitted to Software Impacts).
}
\author{
\strong{Maintainer}: Nicolas Cavan \email{nicolas.cavan@inrae.fr} [contributor]

Authors:
\itemize{
  \item Roland Allart \email{roland.allart@inrae.fr}
  \item Jacques-Eric Bergez \email{jacques-eric.bergez@inrae.fr} (\href{https://orcid.org/0000-0003-3467-2617}{ORCID})
  \item Marta Carpani
  \item Hervé Monod (\href{https://orcid.org/0000-0001-8225-495X}{ORCID})
}

Other contributors:
\itemize{
  \item Aude Alaphilippe (\href{https://orcid.org/0000-0002-4504-2728}{ORCID}) [contributor]
  \item INRAE ((National Research Institute for Agriculture, Food and Environment, France)) [copyright holder]
}

}
\keyword{internal}
