% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov.R
\name{estimate_aov_time}
\alias{estimate_aov_time}
\title{Estimates the execution time for factorial simulations}
\usage{
estimate_aov_time(tree, test_runs = 50)
}
\arguments{
\item{tree}{The Tree object on which simulations are run.}

\item{test_runs}{Number of simulations to be used for time estimation.
Default is 50.}
}
\value{
No explicit return. Prints the estimated execution time.
}
\description{
Estimates the execution time for a specified number of factorial simulations
based on the time taken to run a subset of simulations.
}
