% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synoptique.R
\name{evaluate_and_label}
\alias{evaluate_and_label}
\title{Evaluate Tree Node and Assign Labels}
\usage{
evaluate_and_label(df, tree, option)
}
\arguments{
\item{df}{\code{data.frame} that originates from the tree.}

\item{tree}{\code{Tree} structure object.}

\item{option}{Evaluation options for the tree.}
}
\value{
\code{data.frame} updated with evaluations and labels.
}
\description{
Evaluates each node in the given tree based on the provided options and
then assigns appropriate labels to them.
}
\keyword{internal}
