% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree.R
\name{evaluate_order}
\alias{evaluate_order}
\title{Evaluate Order of Tree Nodes}
\usage{
evaluate_order(tree)
}
\arguments{
\item{tree}{A tree object which contains information about nodes and their
relationships.}
}
\value{
A numeric vector representing the order in which nodes should be
evaluated.
}
\description{
Determines the order in which nodes in a tree should be evaluated, especially
when dealing with aggregated leaves.
}
