% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option.R
\name{evaluate_scenarios}
\alias{evaluate_scenarios}
\title{Evaluate multiple scenarios for a given tree}
\usage{
evaluate_scenarios(aTree, options)
}
\arguments{
\item{aTree}{A "Tree" object}

\item{options}{A matrix representing multiple options to evaluate}
}
\value{
A list of numeric vectors representing the evaluation results of the
scenarios
}
\description{
Works as a wrapper for the \code{evaluate_scenario} function that allows evaluation
of multiple scenarios at once. The scenarios are given as columns in the
\code{options} matrix.
}
