% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oat.R
\name{evaluate_variation}
\alias{evaluate_variation}
\title{Evaluate Variation for Sensitivity Analysis}
\usage{
evaluate_variation(tree, option, leaf_index, variation, range_scale)
}
\arguments{
\item{tree}{A decision tree object to analyze.}

\item{option}{Configuration to be modified for variation.}

\item{leaf_index}{Index of the leaf to be varied.}

\item{variation}{Numeric value indicating the amount and direction of the
variation (+1 or -1).}

\item{range_scale}{The maximum allowed range for the variation.}
}
\value{
Numeric value indicating the evaluation result of the variation or -1
if out of bounds.
}
\description{
Helper function to evaluate a specific variation for the sensitivity
analysis. It checks both positive and negative variations and returns the
evaluation result or -1 if a variation is out of bounds.
}
