% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_subtree.R
\name{extract_names}
\alias{extract_names}
\title{Extract Node Names from Tree Nodes}
\usage{
extract_names(tree_nodes, is_leaf = TRUE)
}
\arguments{
\item{tree_nodes}{a list of \code{Node} objects from which names are to be
extracted.}

\item{is_leaf}{a \code{logical} value indicating whether names should be
extracted from leaf nodes (if \code{TRUE}) or from non-leaf nodes (if
\code{FALSE}).}
}
\value{
A \code{character} vector containing node names.
}
\description{
Retrieves names of nodes based on their leaf status.
}
\keyword{internal}
