% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov.R
\name{generate_aov_formula}
\alias{generate_aov_formula}
\title{Generate AOV Formula}
\usage{
generate_aov_formula(results_df, order = 1)
}
\arguments{
\item{results_df}{Data frame containing AOV results.}

\item{order}{Integer representing the order of interaction; 1 for
first-order, 2 for second-order.}
}
\value{
A formula object for AOV.
}
\description{
Constructs an Analysis of Variance (AOV) formula based on the results
dataframe.
}
