% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synoptique.R
\name{get_box_coordinates}
\alias{get_box_coordinates}
\title{Extract Box Coordinates for Plotting}
\usage{
get_box_coordinates(df)
}
\arguments{
\item{df}{\code{data.frame} with depth and size attributes from the tree.}
}
\value{
\code{data.frame} enhanced with box plotting coordinates, which
includes information about box position, dimensions, and labels.
}
\description{
Extracts the necessary coordinates from the data frame to plot boxes
representing the nodes of a decision tree. The function considers the depth
and size attributes to determine the positions and dimensions of the boxes.
}
\keyword{internal}
