% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree.R
\name{get_chaine}
\alias{get_chaine}
\title{Get the chain of nodes}
\usage{
get_chaine(node_path)
}
\arguments{
\item{node_path}{A character vector representing the path to the node.}
}
\value{
A string representing the chain of nodes
}
\description{
Retrieves the chain of nodes from a given list of nodes.
}
