% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree.R
\name{get_dexi_attributes}
\alias{get_dexi_attributes}
\title{Retrieve Attribute Names from a DEXi XML Tree}
\usage{
get_dexi_attributes(main_tree, root_name)
}
\arguments{
\item{main_tree}{An XML object representing the main tree structure. It must
be a valid XML object containing the DEXi decision model.}

\item{root_name}{A character representing the name of the root attribute from
which to start the search. This should correspond to an existing attribute
name within the XML structure.}
}
\value{
A character vector containing the names of the root attribute and all
its descendant attributes. The vector starts with the name of the root
attribute followed by its child attributes in the order they appear in the
XML tree.
}
\description{
Retrieves all the attribute names under a specified root node within a DEXi
XML tree structure. It forms a search path within the XML tree to fetch all
child attributes starting from the given root name.
}
