% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree.R
\name{get_id}
\alias{get_id}
\title{Retrieve Node ID}
\usage{
get_id(list_nodes, node_name)
}
\arguments{
\item{list_nodes}{A \code{list} of nodes where each node is an object with at
least an \code{Id} and \code{Name} attribute.}

\item{node_name}{\code{character} specifying the name of the desired node.}
}
\value{
\code{numeric} ID of the node if found; \code{NULL} otherwise.
}
\description{
Extracts the ID associated with a specified node name from a provided list of
nodes.
}
\keyword{internal}
