% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree.R
\name{get_leaves}
\alias{get_leaves}
\title{Retrieve leaves associated with a given node in a tree}
\usage{
get_leaves(tree, node_id)
}
\arguments{
\item{tree}{A tree object with nested nodes and leaves.}

\item{node_id}{The identifier of the node for which leaves are to be fetched.}
}
\value{
A vector of leaf IDs associated with the specified node.
}
\description{
Retrieve leaves associated with a given node in a tree
}
