% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_index.R
\name{get_node_id}
\alias{get_node_id}
\title{Retrieve Node ID based on Node Name}
\usage{
get_node_id(tree, node_name, avoid_repetition)
}
\arguments{
\item{tree}{The tree object}

\item{node_name}{The name of the node to find.}

\item{avoid_repetition}{A boolean determining if repeated nodes should be
avoided.}
}
\value{
A numeric value representing the ID of the node.
}
\description{
Retrieves the ID of a specific node from the decision tree based on its name,
considering possible repetitions and the tree structure.
}
