% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree.R
\name{get_paths}
\alias{get_paths}
\title{Retrieve Attribute Paths in an XML Tree}
\usage{
get_paths(attributes, main_tree, root_name)
}
\arguments{
\item{attributes}{\code{character} vector of attribute names whose paths are
to be located. Contains all tree's attributes.}

\item{main_tree}{\code{XML} object representing the tree in which paths are
to be searched.}

\item{root_name}{\code{character} indicating the name of the root attribute
from where the search begins.}
}
\value{
A list of paths where each path is a \code{character} vector,
detailing the sequence of attributes from the root to the desired
attribute.
}
\description{
Navigates through an XML tree to identify specific paths corresponding to a
set of attributes, beginning from a specified root. This function is an
internal tool, chiefly called by \code{\link{create_tree}} to aid in
interpreting and structuring the XML tree.
}
\seealso{
Primary functions that integrate or work with the XML structure:
\itemize{
\item \code{\link{create_tree}}: The main function that invokes
this internal function to dissect the XML output of DEXi.
}
}
\keyword{internal}
