% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_index.R
\name{get_sensitivity_index}
\alias{get_sensitivity_index}
\title{Calculate Sensitivity Index}
\usage{
get_sensitivity_index(tree, avoid_repetition = FALSE)
}
\arguments{
\item{tree}{The tree object for analysis.}

\item{avoid_repetition}{A boolean determining if repeated nodes should be
avoided. Default is FALSE.}
}
\value{
A vector of Sensitivity Indices for each attribute in the tree.
}
\description{
Calculates the Sensitivity Index (SI) of a tree.
}
