% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{getEstimatedWeights}
\alias{getEstimatedWeights}
\alias{getEstimatedWeights.Node}
\title{Get Estimated Weights for Node}
\usage{
getEstimatedWeights(aNode)
}
\arguments{
\item{aNode}{A Node object.}
}
\value{
A numeric vector of estimated weights.
}
\description{
Calculates the estimated weights from the aggregation table of a Node object.
}
