% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_subtree.R
\name{handle_multiple}
\alias{handle_multiple}
\title{Process and Handle Multiple Nodes}
\usage{
handle_multiple(tree, tree_nodes, leaf_names)
}
\arguments{
\item{tree}{a \code{Tree} object representing the original tree.}

\item{tree_nodes}{a list of \code{Node} objects that are currently under
consideration.}

\item{leaf_names}{a \code{character} vector containing the names of leaf
nodes.}
}
\value{
A \code{matrix} detailing occurrences of nodes. If no multiple nodes
are found, a \code{data.frame} is returned instead.
}
\description{
Manages scenarios where nodes might appear multiple times within the tree
structure.
}
\keyword{internal}
