% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo.R
\name{infoMC}
\alias{infoMC}
\title{Estimates the time required to perform a Monte Carlo simulation}
\usage{
infoMC(aTree, nbRuns, iTest = 50)
}
\arguments{
\item{aTree}{Decision tree to run simulations on}

\item{nbRuns}{Total number of Monte Carlo simulations to perform}

\item{iTest}{Number of simulations to be used for time estimation}
}
\value{
No explicit return. The function prints out the estimated execution
time.
}
\description{
Performs a time estimation for a specified number of Monte Carlo simulations.
This estimation is based on the time taken to execute a smaller number of
simulations specified by \code{iTest}.
}
