% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option.R
\name{load_options}
\alias{load_options}
\title{Load options table from a file}
\usage{
load_options(file_name)
}
\arguments{
\item{file_name}{File from which to load the options matrix.}
}
\value{
Matrix of the loaded options.
}
\description{
Retrieves a matrix of options saved in a file. This matrix can then be used
for further analysis or processing.
}
