% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{masc2}
\alias{masc2}
\title{MASC 2.0 model}
\format{
An object of class \code{Tree} of length 1.
}
\source{
\url{https://means-refonte.hub.inrae.fr/outils-emc/masc/telecharger-masc}
}
\usage{
masc2
}
\description{
A Tree class object create by dexiranalysis::create_tree function, using the
Multi-Attribute Assessment of the Sustainability of Cropping systems (MASC)
model in it's 2.0 version.
}
\seealso{
\code{\link{Tree-class}}
}
\keyword{datasets}
