% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonteCarlo.R
\name{MonteCarlo}
\alias{MonteCarlo}
\title{Performs a Monte Carlo simulation}
\usage{
MonteCarlo(aTree, nbRuns, isFile = F, verbose = T)
}
\arguments{
\item{aTree}{Decision tree to run the simulation on}

\item{nbRuns}{Number of Monte Carlo simulations to perform}

\item{isFile}{Logical value to decide whether to write a file or not}

\item{verbose}{Logical value for printing additional information}
}
\value{
Returns a matrix representing the Monte Carlo simulations
}
\description{
Conducts a Monte Carlo simulation on a given decision tree with a specified
number of runs. If \code{isFile} is set to TRUE, a .csv file named "MC
options.csv" is saved, which contains all the random options selected for the
Monte Carlo analysis.
}
