% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\docType{class}
\name{Node-class}
\alias{Node-class}
\title{Title}
\description{
Title
}
\section{Slots}{

\describe{
\item{\code{id}}{(numeric) Unique sequential id of the node.}

\item{\code{name}}{(character) Name of the node.}

\item{\code{isLeaf}}{(logical) Is it a leaf?}

\item{\code{isLeafAndAggregated}}{(logical) Is this leaf also an aggregated node!}

\item{\code{children}}{(character) List of the names of the node's children.}

\item{\code{sisters}}{(character) List of the names of the node's sisters.}

\item{\code{mother}}{(character) Name of the node's mother.}

\item{\code{aggregation}}{(matrix) If aggregated node, table of aggregation.}

\item{\code{Proba}}{(numeric) Estimated weight of aggregation. If Leaf set basically to uniform.}

\item{\code{Depth}}{(numeric) Depth of the node.}

\item{\code{Twin}}{(numeric) In case of multiple leaves, give the id of the other leaves.}

\item{\code{CondiProbaList}}{(list) ???????????}

\item{\code{rangeScale}}{(numeric) Range scale.}

\item{\code{scaleLabel}}{(character) Labels of the different scales.}

\item{\code{nodePath}}{(character) Node path from root to leaf.}
}}

