% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OAT.R
\name{OAT}
\alias{OAT}
\title{Perform One-Factor-At-A-Time (OFAT) Sensitivity Analysis}
\usage{
OAT(aTree, option)
}
\arguments{
\item{aTree}{A decision tree object to perform the analysis on.}

\item{option}{The initial configuration for the decision tree parameters.}
}
\value{
A matrix with evaluation results for each attribute in the decision
tree, under different variations of the parameters.
}
\description{
Conducts One-Factor-At-A-Time (OFAT) sensitivity analysis by performing
simulations with varying factors, keeping all but one factor constant for
each simulation.
}
