% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_subtree.R
\name{obtain_attribute_ids}
\alias{obtain_attribute_ids}
\title{Identify Descendant Attribute IDs of a Node}
\usage{
obtain_attribute_ids(tree, node_id)
}
\arguments{
\item{tree}{a \code{Tree} object.}

\item{node_id}{a \code{numeric} or \code{character} identifier representing
the \code{Node} that will act as the reference point.}
}
\value{
A \code{numeric} vector of attribute IDs symbolizing descendant
nodes.
}
\description{
Retrieves all nodes that share a path with the specified node. Such nodes are
considered descendants of the given node and are included in the subtree.
}
\keyword{internal}
