% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_index.R
\name{plot_sensitivity_index}
\alias{plot_sensitivity_index}
\title{Show Sensitivity Index (SI)}
\usage{
plot_sensitivity_index(tree, sensitivity_indices)
}
\arguments{
\item{tree}{\code{Tree} object for analysis.}

\item{sensitivity_indices}{A \code{numeric} vector representing the
Sensitivity Indices for each leaf in the decision tree.}
}
\value{
No return value; a bar plot is displayed.
}
\description{
Visualizes the Sensitivity Index (SI) of the leaves of a specified decision
tree using a bar plot.
}
\examples{
tree <- dexiranalysis::masc2
sensitivity_index <- si_dexi(tree)
plot_sensitivity_index(tree, sensitivity_index[[1]])

}
