% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{print,Node-method}
\alias{print,Node-method}
\alias{print.Node}
\title{print method for Node class object}
\usage{
\S4method{print}{Node}(x, ...)
}
\arguments{
\item{x}{The Node object to be printed.}

\item{...}{additional parameters to be passed to the print function.}
}
\description{
Prints basic information about the node including its name, id, depth, path, and other properties.
}
