% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Node.R
\name{print,Node-method}
\alias{print,Node-method}
\title{print.node}
\usage{
\S4method{print}{Node}(x, ...)
}
\arguments{
\item{x}{The printed Node}

\item{...}{see '...' print prameters}
}
\description{
Print method for Node class object
}
