% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{print,Tree-method}
\alias{print,Tree-method}
\alias{print.Tree}
\alias{\S4method{print}{Tree}}
\title{Print Method for Tree Class Objects}
\usage{
\S4method{print}{Tree}(x, ...)
}
\arguments{
\item{x}{An object of class \code{Tree} that you want to print.}

\item{...}{Additional arguments to be passed to the underlying print
function, though they might not have any effect in this custom print
method.}
}
\value{
This function is invoked for its side effect of printing. It does not
return anything.
}
\description{
Custom print method for objects of class \code{Tree}. This method prints out
basic information about the tree, such as the root name, number of
attributes, number of leaves, depth, and details about specific nodes.
}
\seealso{
\itemize{
\item \code{\link{Tree-class}}: For the Tree class definition.
\item \code{\link{show.Tree}}: For showing a Tree object.
\item \code{\link{describe.Tree}}: For describing a Tree object.
}
}
