% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{print,Tree-method}
\alias{print,Tree-method}
\alias{print.Tree}
\title{print method for Tree class object}
\usage{
\S4method{print}{Tree}(x, ...)
}
\arguments{
\item{x}{Tree object to be printed.}

\item{...}{additional parameters to be passed to the print function.}
}
\description{
Custom print method for Tree class object. Prints basic information about the
tree including root name, number of attributes, leaves, depth and specific
nodes.
}
