% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree.R
\name{process_duplicated_leaf}
\alias{process_duplicated_leaf}
\title{Process Duplicated leaf}
\usage{
process_duplicated_leaf(tree_nodes, leaves)
}
\arguments{
\item{tree_nodes}{A list of nodes in the tree structure.}

\item{leaves}{A character vector containing the names of leaf attributes.}
}
\value{
A list containing two elements:
\itemize{
\item \code{multiple}: A data frame or matrix containing the count of
occurrences of leaf attributes.
\item \code{is_multiple}: A logical value indicating whether there are multiple
occurrences of any leaf attributes.
\item \code{tree_nodes}: A list of tree nodes updated with the twin
information and markings for duplicated leaf.
}
}
\description{
Identifies the multiple occurrences of leaf attributes in the tree structure,
marking them with the appropriate \code{Twin} reference. It also returns a data
frame containing the count of occurrences.
}
