% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option.R
\name{save_options}
\alias{save_options}
\title{Save Options Table}
\usage{
save_options(options_table, file_name)
}
\arguments{
\item{options_table}{A \code{matrix} containing option values.}

\item{file_name}{A \code{character} string specifying the desired file name for
saving the options.}
}
\value{
No return value, called for side effects
}
\description{
Stores a matrix of options into a file, primarily for archival or subsequent
analysis.
}
\examples{
tree <- dexisensitivity::masc2
option <- create_options(tree, num_options=3, seed = 42)
save_options(option,paste0(tempdir(),"\\\\save_options.tab"))
file.remove(paste0(tempdir(),"\\\\save_options.tab"))

}
