% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AOV.R
\name{sensib.effet}
\alias{sensib.effet}
\title{Title}
\usage{
sensib.effet(aov.obj, nb.plot = 8, beside = T, las = 1, ...)
}
\arguments{
\item{aov.obj}{objet "aov" issu de l'analyse de variance des donnees de simulation}

\item{nb.plot}{nombre d'effets repr?sent?s sur le graphique}

\item{beside}{voir la doc de barplot}

\item{las}{las}

\item{...}{...}
}
\description{
Fonction de calcul des indices de sensibilite factoriels
}
