% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AOV.R
\name{sensib.grpe}
\alias{sensib.grpe}
\title{Title}
\usage{
sensib.grpe(aov.obj, fact.interet, las = 1, ...)
}
\arguments{
\item{aov.obj}{objet "aov" issu de l'analyse de variance des donnees de simulation}

\item{fact.interet}{fact.interet}

\item{las}{las}

\item{...}{...}
}
\description{
Fonction de calcul des indices de sensibilite factoriels
}
