% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AOV.R
\name{sensib.total}
\alias{sensib.total}
\title{Title}
\usage{
sensib.total(aov.obj, main.show = T, las = 1, ...)
}
\arguments{
\item{aov.obj}{objet "aov" issu de l'analyse de variance des donnees de simulation}

\item{main.show}{main.show}

\item{las}{las}

\item{...}{...}
}
\description{
Fonction de calcul des indices de sensibilite factoriels
}
