% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensitivityIndex.R
\name{sensitivity.condprob}
\alias{sensitivity.condprob}
\title{Calculate Sensitivity Index (SI) From Conditional Probabilities}
\usage{
sensitivity.condprob(condproblist, weightlist)
}
\arguments{
\item{condproblist}{A list of matrices of conditional probabilities (Y conditional to each A_i) plus the vector of Y probabilities.}

\item{weightlist}{The list of weights of the A_i factor levels.}
}
\value{
A vector of Sensitivity Indices (SI).
}
\description{
Calculates the first order sensitivity indices of Y with respect to A_i descendants, when given the conditional probabilities and A_i weights.
}
