% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Options.R
\name{showScenario}
\alias{showScenario}
\title{Plot a bar chart of a single scenario}
\usage{
showScenario(aScenario, aTree, isLabelY = TRUE, isPar = T)
}
\arguments{
\item{aScenario}{the scenario to graph}

\item{aTree}{The associated "Tree" object}

\item{isLabelY}{A logical value indicating whether to include labels on the Y axis (Default: TRUE)}

\item{isPar}{A logical value indicating whether to modify the graph's parameters (Default: TRUE)}
}
\description{
Visualizes the values assigned to each attribute in a given scenario
and marks the maximum possible value for each attribute using a bar chart.
}
\seealso{
\code{\link{EvaluateScenario}}
}
