% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensitivityIndex.R
\name{SI_DEXi}
\alias{SI_DEXi}
\title{Title}
\usage{
SI_DEXi(aTree, fileName = "SI_out.csv", isFile = T, avoidrep = F)
}
\arguments{
\item{aTree}{aTree}

\item{fileName}{fileName}

\item{isFile}{isFile}

\item{avoidrep}{T if you want to avoid repeted nodes}
}
\description{
Calculation of the Sensitivity Index (SI)
}
