% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\docType{class}
\name{Tree-class}
\alias{Tree-class}
\title{Tree}
\description{
Tree class definition
}
\section{Slots}{

\describe{
\item{\code{nbAttributes}}{(numeric) Number of attributes}

\item{\code{nbLeaves}}{(numeric) Number of leaves}

\item{\code{Depth}}{(numeric) Maximum depth of the tree}

\item{\code{Attributes}}{(character) List of names of attributes}

\item{\code{Leaves}}{(character) List of names of leaves}

\item{\code{Aggregated}}{(character) List of names of aggregated nodes}

\item{\code{isMultiple}}{(logical) Tag to know if multiple leaves}

\item{\code{Multiple}}{(data.frame) List of the multiple leaves and number of occurrence}

\item{\code{isLeafAggregated}}{(logical) Tag to know if leaf-aggregated nodes}

\item{\code{LeafAggregated}}{(character) List of names of leaf-aggregated nodes}

\item{\code{Paths}}{(list) Path from root to leafs}

\item{\code{Nodes}}{(list) List of nodes}

\item{\code{EvalOrder}}{(numeric) Evaluation order in case of LeafAggregated nodes}

\item{\code{rootName}}{(character) Name of the root node}
}}

