% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\docType{class}
\name{Tree-class}
\alias{Tree-class}
\alias{Tree}
\title{An S4 class to represent a Tree}
\value{
An object of class Tree.
}
\description{
It includes several slots to store attributes, leaves, nodes, path, and
additional information related to the tree structure.
}
\section{Slots}{

\describe{
\item{\code{NumberOfAttributes}}{Object of class "numeric", the number of attributes in the
tree.}

\item{\code{NumberOfLeaves}}{Object of class "numeric", the number of leaves in the tree.}

\item{\code{Depth}}{Object of class "numeric", the maximum depth of the tree.}

\item{\code{Attributes}}{Object of class "character", list of names of attributes in
the tree.}

\item{\code{Leaves}}{Object of class "character", list of names of leaves in the
tree.}

\item{\code{Aggregated}}{Object of class "character", list of names of aggregated
nodes in the tree.}

\item{\code{IsMultiple}}{Object of class "logical", flag indicating if there are
multiple leaves in the tree.}

\item{\code{Multiple}}{Object of class "data.frame", list of multiple leaves and
their count.}

\item{\code{IsLeafAggregated}}{Object of class "logical", flag indicating if there
are leaf-aggregated nodes in the tree.}

\item{\code{LeafAggregated}}{Object of class "character", list of names of
leaf-aggregated nodes in the tree.}

\item{\code{Paths}}{Object of class "list", list of paths from the root to the
leaves.}

\item{\code{Nodes}}{Object of class "list", list of nodes in the tree.}

\item{\code{EvaluationOrder}}{Object of class "numeric", evaluation order in case of
LeafAggregated nodes.}

\item{\code{RootName}}{Object of class "character", name of the root node.}
}}

\seealso{
\code{\link{print.Tree}}, \code{\link{show.Tree}},
\code{\link{describe.Tree}}
}
